### 2024-05-27 #6793 @editorialbot
[PRE REVIEW]: t8code - modular adaptive mesh refinement in the exascale era
**Submitting author:** @jmark (<a href="http://orcid.org/0000-0001-6297-9494">Johannes Markert</a>)
**Repository:** https://github.com/DLR-AMR/t8code
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v2.0.0
