### 2024-05-27 #6793 @danielskatz
@openjournals/dev - Do you have any idea why this didn't work automatically?
