### 2024-05-27 #6793 @danielskatz
Also, please fix the one invalid DOI in your .bib file.
