### 2024-05-27 b3f3148
Use t8code logo, exclude source-file documentation
