+++ JOSS/paper.bib
@@ -446,7 +446,7 @@ volume = {245},
 pages = {106866},
 year = {2019},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2019.106866},
+doi = {10.1016/j.cpc.2019.106866},
 url = {https://www.sciencedirect.com/science/article/pii/S001046551930253X},
 author = {J. Teunissen and R. Keppens},
 keywords = {Multigrid, Elliptic solver, Octree, Adaptive mesh refinement, Divergence cleaning},
