### 2024-05-27 #6793 @jmark
> Also, please fix the one invalid DOI in your .bib file.

Done. Let me know if I can be of any help!
