### 2024-05-27 #6793 @elbeejay
@jmark I'm going to be handling this paper as we work in through the JOSS review process. If you have any suggestions for potential reviewers, please comment with their names and GitHub handle's here (omit the '@' so they don't get a notification unnecessarily).

Thanks! Otherwise I'll have a look and start contacting potential reviewers this week.
