### 2024-05-29 6acb26d
Make compile mode cyclable.
