### 2024-05-29 #6793 @elbeejay
@rmrsk would you be willing to review this JOSS submission, "t8code - modular adaptive mesh refinement in the exascale era"? You have the background of a great reviewer for this software, which is written in C/C++ and presents a modular approach to adaptive mesh refinement.

I understand you've published with JOSS, so I expect you're familiar with our open approach to reviewing. Please let us know if you are able and willing to review.

If you cannot provide your expertise as a reviewer, please take a minute and let us know if you have any colleagues, peers, or trainees who might be suitable reviewers.

Thanks,
Jay
