### 2024-06-03 #6793 @elbeejay
Hi @DamynChipman, would you be interested in reviewing this submission "t8code - modular adaptive mesh refinement in the exascale era", for JOSS? I saw you've recently published your C++ software `EllipticForest` with JOSS, and thought you might have the expertise in both the programming langage and the scientific domain necessary to serve as a reviewer for this submission.

Let me know if you are interested and would have time to provide your review.

Thanks!
Jay
