### 2024-06-04 #6793 @sloede
> @sloede would you be interested in reviewing this submission to JOSS titled "t8code - modular adaptive mesh refinement in the exascale era"?

Dear @elbeejay thanks for the invite. I'd love to review this paper as it is right up my alley, but unfortunately I have to excuse myself due to a conflict of interest. Feel free to ping me again if a similar publication should pop up in the future though.
