### 2024-06-04 925e56d
Add const modifier to t8_element_array
