### 2024-06-04 b996034
Add declaration and documentation of new functions
