### 2024-06-04 #6793 @DamynChipman
Hi @elbeejay, I would be happy to review this paper. I have a question regarding COI: While not one of the authors, I am currently collaborating with Carsten Burstedde on separate work. Carsten has made contributions to the code. Does that qualify as a COI?
