### 2024-06-04 #6793 @danielskatz
@DamynChipman - This isn't a formal COI, though it makes sense for you to mention it here. In my opinion (as the track chair), if you feel you can review this fairly, I think it would be ok for you to be a reviewer.
