+++ src/t8_cmesh/t8_cmesh_examples.h
@@ -133,6 +133,50 @@ t8_cmesh_t
 t8_cmesh_new_hypercube_pad (const t8_eclass_t eclass, sc_MPI_Comm comm, const double *boundary, t8_locidx_t polygons_x,
                             t8_locidx_t polygons_y, t8_locidx_t polygons_z, const int use_axis_aligned);
 
+/** Construct a hypercube forest from one primitive tree class.
+ * \param [in] eclass       This element class determines the dimension of the cube.
+ * \param [in] comm         The mpi communicator to be used.
+ * \param [in] boundary     The vertices, that define the hypercube boundary.
+ * \param [in] polygons_x   The number of polygons along the x-axis.
+ * \param [in] polygons_y   The number of polygons along the y-axis.
+ *                              Only required if \a eclass is 2D or 3D.
+ * \param [in] polygons_z       The number of polygons along the z-axis.
+ *                              Only required if \a eclass is 3D.
+ * \param [in] periodic_x   Connect opposite sides of the hypercube in x-direction.
+ * \param [in] periodic_y   Connect opposite sides of the hypercube in y-direction.
+ * \param [in] periodic_z   Connect opposite sides of the hypercube in z-direction.
+ * \param [in] use_axis_aligned Use the axis-aligned geometry. If used, only two points per tree are stored.
+ * \param [in] set_partition  If true, partition the cmesh.
+ * \param [in] offset         Offset of the local tree ids for a given partition.
+ * \return                      A committed t8_cmesh structure with 
+ *                              \a polygons_x * \a polygons_z * \a polygons_y many 
+ *                              sub-hypercubes of class \a eclass.
+ * \note \a boundary must point to an array with 3*8 (3D), 3*4 (2D), 3*2 (1D), or 3 (0D) entries.
+ * \note Every sub-hypercube contains different number of trees depending on \a eclass.
+ * \note If \a eclass == T8_ECLASS_VERTEX, _LINE, _QUAD or _HEX every sub-hypercube contains
+ *  one tree, if _TRIANGLE or _PRISM two trees and if _TET six trees.
+ *  This is done in the same way as in \see t8_cmesh_new_hypercube.
+ * \example let eclass = T8_ECLASS_TRIANGLE
+ *              boundary coordinates = a(0,0,0), b(3,0,0), c(0,2,0), d(3,2,0)
+ *              polygons_x, _y, _z = 3, 1, 0                 
+ *      
+ *    c--f--h--d     The hypercube defined by the boundary coordinates
+ *    |  |  |  |     is first split into 3 sub-hypercubes. The sub-hypercubes
+ *    |  |  |  |     are ordered from left to right (and top to bottom).
+ *    a--e--g--b     Coordinates e,f,g,h are (1,0,0),(1,2,0),(2,0,0),(2,2,0).
+ * 
+ *    c--f--h--d     Each sub-hypercube is the split into 2 triangle roots.
+ *    |1/|3/|5/|     The ordering is the same as in \see t8_cmesh_new_hypercube.
+ *    |/0|/2|/4|     Thus, we get 6 trees, which are ordered as shown in the picture. 
+ *    a--e--g--b     
+ *
+ */
+t8_cmesh_t
+t8_cmesh_new_hypercube_pad_ext (const t8_eclass_t eclass, sc_MPI_Comm comm, const double *boundary,
+                                t8_locidx_t polygons_x, t8_locidx_t polygons_y, t8_locidx_t polygons_z,
+                                const int periodic_x, const int periodic_y, const int periodic_z,
+                                const int use_axis_aligned, const int set_partition, t8_gloidx_t offset);
+
 /** Hybercube with 6 Tets, 6 Prism, 4 Hex. 
[…]
