### 2024-06-05 12c5abd
Add declaration of is_element and tree_is_local
