+++ src/t8_forest/t8_forest_general.h
@@ -436,6 +436,17 @@ t8_forest_get_num_ghosts (const t8_forest_t forest);
 t8_eclass_t
 t8_forest_get_eclass (const t8_forest_t forest, const t8_locidx_t ltreeid);
 
+/**
+ * Check whether a given tree id belongs to a local tree in a forest.
+ * 
+ * \param [in]    forest The forest.
+ * \param [in]    local_tree A tree id.
+ * \return True if and only if the id \a local_tree belongs to a local tree of \a forest.
+ * \a forest must be committed before calling this function.
+ */
+int
+t8_forest_tree_is_local (const t8_forest_t forest, const t8_locidx_t local_tree);
+
 /** Given a global tree id compute the forest local id of this tree.
  * If the tree is a local tree, then the local id is between 0 and the number
  * of local trees. If the tree is not a local tree, a negative number is returned.
@@ -492,6 +503,19 @@ t8_forest_cmesh_ltreeid_to_ltreeid (t8_forest_t forest, t8_locidx_t lctreeid);
  */
 t8_ctree_t
 t8_forest_get_coarse_tree (t8_forest_t forest, t8_locidx_t ltreeid);
+
+/**
+ * Query whether a given element is a leaf in a forest.
+ * 
+ * \param [in]  forest    The forest.
+ * \param [in]  element   An element of a local tree in \a forest.
+ * \param [in]  local_tree A local tree id of \a forest.
+ * \return True (non-zero) if and only if \a element is a leaf in \a local_tree of \a forest.
+ * \note This does not query for ghost leafs.
+ * \note \a forest must be committed before calling this function.
+ */
+int
+t8_forest_element_is_leaf (const t8_forest_t forest, const t8_element_t *element, const t8_locidx_t local_tree);
 
 /** Compute the leaf face orientation at given face in a forest.
  * \param [in]    forest  The forest. Must have a valid ghost layer.
