### 2024-06-08 #6793 @elbeejay
Hi @jmsexton03 and @asalmgren would either of you be interested in reviewing this submission "t8code - modular adaptive mesh refinement in the exascale era" for JOSS? From what I can tell, the software you published with JOSS in 2021, Nyx, solves cosmological hydrodynamics over adaptive meshes, which made me think you might be suitable reviewers for this package on adaptive mesh refinement. Please let me know if you are interested or if you have any further questions, thanks!

Jay
