+++ JOSS/paper.md
@@ -55,10 +55,12 @@ authors:
     affiliation: 1
   - name: Carsten Burstedde
     orcid: 0000-0001-9843-1041
-    affiliation: 1
+    affiliation: 2
 affiliations:
  - name: German Aerospace Center (DLR), Institute for Software Technology, Cologne, Germany
    index: 1
+ - name: Rheinische Friedrich-Wilhelms-Universität Bonn, Institute for Numerical Simulations and Hausdorff Center for Mathematics,  Germany
+   index: 2
 date: 10 June 2024
 bibliography: paper.bib
 
