### 2024-06-11 #6793 @elbeejay
Hi @ctlee - I'm reaching out to ask if you would consider reviewing the submission "t8code - modular adaptive mesh refinement in the exascale era" for The Journal of Open Source Software (JOSS). I saw your work on the [gamer](https://github.com/ctlee/gamer) project and thought you might be a good reviewer for this submission. I believe you have some exposure to JOSS, as a co-author on the ["SMART: Spatial Modeling Algorithms for Reactions and Transport"](https://joss.theoj.org/papers/10.21105/joss.05580) paper, but I will provide a short summary of JOSS below just in case.

At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html). This issue is a "pre-review issue" which we use to find peer-reviewers. Once 3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know!

Thanks,
Jay
