### 2024-06-12 c035cb5
Better documentaion in t8_forest_transition.h
