+++ src/t8_forest/t8_forest_transition.h
@@ -36,7 +36,8 @@
 
 T8_EXTERN_C_BEGIN ();
 
-/* In this function, use a binary encoding (depending on the face enumeration), to determine which subelement type to use. 
+
+ /**  In this function, use a binary encoding (depending on the face enumeration), to determine which subelement type to use. 
  * Every face has a flag parameter, which is set to 1, if there is a neighbour with a higher level 
  * and to 0, if the level of the neighbour is at most the level of the element.   
  *             
@@ -53,29 +54,58 @@ T8_EXTERN_C_BEGIN ();
  * Note, that this procedure is independent of the eclass (we only show an example for the quad scheme). 
  * Each neighbour-structure will lead to a unique binary code. 
  * Within the element scheme of the given eclass, this binary code is used to construct the right subelement type,
- * in order to remove hanging nodes from the mesh. */
-int                 t8_forest_transition_adapt (t8_forest_t forest,
-                                                t8_forest_t
-                                                forest_from,
-                                                t8_locidx_t
-                                                ltree_id,
-                                                t8_locidx_t
-                                                lelement_id,
-                                                t8_eclass_scheme_c *ts,
-                                                int num_elements,
-                                                t8_element_t *elements[]);
+ * in order to remove hanging nodes from the mesh.
 
-/* This function is the point of entry for the hanging-faces-removing subelements. 
+ * \param [in] forest       the forest to which the new elements belong
+ * \param [in] forest_from  the forest that is adapted.
+ * \param [in] which_tree   the local tree containing \a elements
+ * \param [in] lelement_id  the local element id in \a forest_old in the tree of the current element
+ * \param [in] ts           the eclass scheme of the tree
+ * \param [in] is_family    if 1, the first \a num_elements entries in \a elements form a family. If 0, they do not.
+ * \param [in] num_elements the number of entries in \a elements that are defined
+ * \param [in] elements     Pointers to a family or, if \a is_family is zero,
+ *                          pointer to one element.
+ * \return transition type + 1 (to avoid return value 1)
+ */
+
+
+int
+t8_forest_transition_adapt (t8_forest_t forest,
+                            t8_forest_t forest_from,
+                            t8_locidx_t ltree_id, 
+                            t8_locidx_t lelement_id,
+                            t8_eclass_scheme_c *ts,
+                            int num_elements,
+                            t8_element_t *elements[]);
+
+
+ /** This function is the point of entry for the hanging-faces-removing subelements. 
  * In this function, the corresponding callback function for the refine value in forest_adapt is set
- * and the forest is adapted. */
-void                t8_forest_transition (t8_forest_t forest);
+ * and the forest is adapted.
+ * \param [in,out] forest       The input forest consists of hanging nodes on faces- 
+ *                              The output forest contains transition cells and no hanging nodes on faces. 
+ */
+void
+t8_forest_transition (t8_forest_t forest);
 
-/* This function is the point of entry in order to untransition a forest before adapting it. 
- * Every transition cell gets coarsened back to its parent element. */
-void                t8_forest_untransition (t8_forest_t forest);
 
-/* Check whether the forest is transitioned, meaning that subelements exist. */
-int                 t8_forest_is_transitioned (t8_forest_t forest);
+/** This function is the point of entry in order to untransition a forest before adapting it. 
+ * Every transition cell gets coarsened back to its parent element.
+ * \param [in,out] forest       The input forest consists of subelements- 
+ *                              The output forest does not. 
+ */
+
+void
+t8_forest_untransition (t8_forest_t forest);
+
+ /** Check whether the forest is transitioned, meaning that subelements exist. 
[…]
