+++ src/CMakeLists.txt
@@ -206,12 +206,29 @@ set( T8_PUBLIC_HEADERS
 )
 
 install( FILES ${T8_PUBLIC_HEADERS} DESTINATION ${CMAKE_INSTALL_PREFIX}/include )
-install( TARGETS T8
-  EXPORT T8CODEConfig
-  DESTINATION ${CMAKE_INSTALL_PREFIX}/lib
+install( TARGETS T8 DESTINATION ${CMAKE_INSTALL_PREFIX}/lib )
+install( TARGETS T8 EXPORT ${PROJECT_NAME}-targets )
+
+include( CMakePackageConfigHelpers )
+
+configure_package_config_file( ${CMAKE_CURRENT_LIST_DIR}/config.cmake.in
+  ${CMAKE_CURRENT_BINARY_DIR}/CMakeFiles/${PROJECT_NAME}Config.cmake
+INSTALL_DESTINATION cmake
 )
 
-install( EXPORT T8CODEConfig
-  DESTINATION "${CMAKE_INSTALL_PREFIX}/cmake"
-  NAMESPACE T8CODE::
+write_basic_package_version_file(
+${CMAKE_CURRENT_BINARY_DIR}/CMakeFiles/${PROJECT_NAME}ConfigVersion.cmake
+COMPATIBILITY SameMajorVersion
+VERSION 0.0.0
+)
+
+install( EXPORT ${PROJECT_NAME}-targets
+NAMESPACE ${PROJECT_NAME}::
+DESTINATION cmake
+)
+
+install( FILES
+${CMAKE_CURRENT_BINARY_DIR}/CMakeFiles/${PROJECT_NAME}Config.cmake
+${CMAKE_CURRENT_BINARY_DIR}/CMakeFiles/${PROJECT_NAME}ConfigVersion.cmake
+DESTINATION cmake
 )
