+++ src/t8_forest/t8_forest.c
@@ -78,28 +78,25 @@ t8_forest_is_initialized (t8_forest_t forest)
 int
 t8_forest_supports_transitioning (t8_forest_t forest)
 {
-  int                 supports_transition = 1;
-  int                 supports_transition_all_procs = 0;        /* Result over all procs */
-  int                 int_eclass;
-  int                 mpiret;
+  int supports_transition = 1;
+  int supports_transition_all_procs = 0; /* Result over all procs */
+  int int_eclass;
+  int mpiret;
   t8_eclass_scheme_c *tscheme;
 
   /* Iterate over all eclasses */
-  for (int_eclass = (int) T8_ECLASS_ZERO; int_eclass < (int) T8_ECLASS_COUNT;
-       int_eclass++) {
+  for (int_eclass = (int) T8_ECLASS_ZERO; int_eclass < (int) T8_ECLASS_COUNT; int_eclass++) {
     /* If the forest has trees of the current eclass, check if elements of this
      * eclass supports transitioning. */
     if (forest->cmesh->num_local_trees_per_eclass[int_eclass] > 0) {
       tscheme = forest->scheme_cxx->eclass_schemes[int_eclass];
-      supports_transition = supports_transition
-        && t8_element_scheme_supports_transitioning (tscheme);
+      supports_transition = supports_transition && t8_element_scheme_supports_transitioning (tscheme);
     }
   }
   /* Combine the process-local results via a logic or and distribute the
    * result over all procs (in the communicator).*/
-  mpiret =
-    sc_MPI_Allreduce (&supports_transition, &supports_transition_all_procs, 1,
-                      sc_MPI_INT, sc_MPI_LAND, forest->mpicomm);
+  mpiret = sc_MPI_Allreduce (&supports_transition, &supports_transition_all_procs, 1, sc_MPI_INT, sc_MPI_LAND,
+                             forest->mpicomm);
   SC_CHECK_MPI (mpiret);
 
   return supports_transition_all_procs;
@@ -256,17 +253,15 @@ t8_forest_set_balance (t8_forest_t forest, const t8_forest_t set_from, int no_re
 }
 
 void
-t8_forest_set_transition (t8_forest_t forest, const t8_forest_t set_from,
-                          int set_transition_with_balance)
+t8_forest_set_transition (t8_forest_t forest, const t8_forest_t set_from, int set_transition_with_balance)
 {
   T8_ASSERT (t8_forest_is_initialized (forest));
 
   if (set_transition_with_balance) {
-    t8_debugf("-------------set transition with balance -------------\n");
+    t8_debugf ("-------------set transition with balance -------------\n");
     /* balance with repartition */
     t8_forest_set_balance (forest, set_from, 0);
   }
-
 
   if (set_from != NULL) {
     /* Note that it is possible to apply transitioning to a forest without transition implementation.
@@ -546,8 +541,8 @@ t8_forest_commit (t8_forest_t forest)
     forest->incomplete_trees = 0;
     forest->is_transitioned = 0;
   }
-  else { 
-                                   /* set_from != NULL */
+  else {
+    /* set_from != NULL */
     t8_forest_t forest_from = forest->set_from; /* temporarily store set_from, since we may overwrite it */
 
     T8_ASSERT (forest->mpicomm == sc_MPI_COMM_NULL);
@@ -600,18 +595,18 @@ t8_forest_commit (t8_forest_t forest)
       if (forest->from_method > 0) {
         /* The forest should also be partitioned/balanced/transitioned.
[…]
