### 2024-06-12 #6793 @asalmgren
Unfortunately I am slammed this summer and wouldn't have time to do a
competent review in a reasonable amount of time.

My apologies.

Ann

On Sat, Jun 8, 2024 at 3:44 AM J. Hariharan ***@***.***>
wrote:

> Hi @jmsexton03 <https://github.com/jmsexton03> and @asalmgren
> <https://github.com/asalmgren> would either of you be interested in
> reviewing this submission "t8code - modular adaptive mesh refinement in the
> exascale era" for JOSS? From what I can tell, the software you published
> with JOSS in 2021, Nyx, solves cosmological hydrodynamics over adaptive
> meshes, which made me think you might be suitable reviewers for this
> package on adaptive mesh refinement. Please let me know if you are
> interested or if you have any further questions, thanks!
>
> Jay
>
