### 2024-06-13 867c768
Add the exhale extension to sphinx for a better structured documentation
