### 2024-06-14 #6887 @elbeejay
Hi @jwallwork23, @DamynChipman, and @alecontri - thank you all again for offering to peer-review `t8code`. Hopefully the peer review process ultimately results in the creation of a software tool that is useful to the broader community.

Please see the instructions at the top of this issue on how to generate your reviewer checklist and a link to the reviewer guidelines. Do not hesitate to reach out to me if you have any questions about the JOSS review process.

Right now we are asking that reviewers try to complete their reviews in 6 weeks if possible, so by July 25 in this case. Please keep us posted here if you think you will need an extension. I will ask the bot to issue reminders in 3 weeks so we all stay aware of this on-going review.

Thanks again,
Jay
