### 2024-06-18 2517ecd
Cut folder path
