### 2024-06-18 37c64cc
Do not allow offsets to compute negative num_trees
