+++ src/t8_cmesh/t8_cmesh_offset.c
@@ -64,8 +64,9 @@ t8_offset_num_trees (const int proc, const t8_gloidx_t *offset)
   T8_ASSERT (offset != NULL);
 
   num_global_trees = T8_GLOIDX_ABS (offset[proc + 1]) - t8_offset_first (proc, offset);
+  T8_ASSERT (num_global_trees >= 0);
   /* If num_global_trees < 0, we return 0 */
-  return num_global_trees >= 0 ? num_global_trees : 0;
+  return num_global_trees;
 }
 
 /* The last local tree of a given process in a partition */
