### 2024-06-18 6621b16
replace global by local
