+++ src/t8_cmesh/t8_cmesh_offset.c
@@ -59,14 +59,14 @@ t8_offset_first_tree_to_entry (const t8_gloidx_t first_tree, const int shared)
 t8_gloidx_t
 t8_offset_num_trees (const int proc, const t8_gloidx_t *offset)
 {
-  t8_gloidx_t num_global_trees;
+  t8_gloidx_t num_local_trees;
   T8_ASSERT (proc >= 0);
   T8_ASSERT (offset != NULL);
 
-  num_global_trees = T8_GLOIDX_ABS (offset[proc + 1]) - t8_offset_first (proc, offset);
-  T8_ASSERT (num_global_trees >= 0);
+  num_local_trees = T8_GLOIDX_ABS (offset[proc + 1]) - t8_offset_first (proc, offset);
+  T8_ASSERT (num_local_trees >= 0);
   /* If num_global_trees < 0, we return 0 */
-  return num_global_trees;
+  return num_local_trees;
 }
 
 /* The last local tree of a given process in a partition */
