### 2024-06-18 87f89fd
start fixing cmesh copy
