+++ src/t8_cmesh/t8_cmesh_copy.c
@@ -69,13 +69,15 @@ t8_cmesh_copy (t8_cmesh_t cmesh, t8_cmesh_t cmesh_from, sc_MPI_Comm comm)
           T8_ECLASS_COUNT * sizeof (t8_locidx_t));
 
   /* Copy the tree info */
-  num_parts = t8_cmesh_trees_get_numproc (cmesh_from->trees);
   cmesh->trees = NULL;
-  t8_cmesh_trees_init (&cmesh->trees, num_parts, cmesh_from->num_local_trees, cmesh_from->num_ghosts);
-  t8_cmesh_trees_copy_toproc (cmesh->trees, cmesh_from->trees, cmesh_from->num_local_trees, cmesh_from->num_ghosts);
-  for (iz = 0; iz < num_parts; iz++) {
-    t8_cmesh_trees_get_part_data (cmesh_from->trees, iz, &first_tree, &num_trees, &first_ghost, &num_ghosts);
-    t8_cmesh_trees_start_part (cmesh->trees, iz, first_tree, num_trees, first_ghost, num_ghosts, 0);
-    t8_cmesh_trees_copy_part (cmesh->trees, iz, cmesh_from->trees, iz);
+  if (cmesh_from->trees) {
+    num_parts = t8_cmesh_trees_get_numproc (cmesh_from->trees);
+    t8_cmesh_trees_init (&cmesh->trees, num_parts, cmesh_from->num_local_trees, cmesh_from->num_ghosts);
+    t8_cmesh_trees_copy_toproc (cmesh->trees, cmesh_from->trees, cmesh_from->num_local_trees, cmesh_from->num_ghosts);
+    for (iz = 0; iz < num_parts; iz++) {
+      t8_cmesh_trees_get_part_data (cmesh_from->trees, iz, &first_tree, &num_trees, &first_ghost, &num_ghosts);
+      t8_cmesh_trees_start_part (cmesh->trees, iz, first_tree, num_trees, first_ghost, num_ghosts, 0);
+      t8_cmesh_trees_copy_part (cmesh->trees, iz, cmesh_from->trees, iz);
+    }
   }
 }
