### 2024-06-19 d4cd1f0
add is_leaf assertions
