### 2024-06-19 4f1755a
proper const and mutable version of t8_element_array_get_data_mutable
