+++ src/t8_forest/t8_forest.c
@@ -945,7 +945,7 @@ t8_forest_get_element_in_tree (t8_forest_t forest, t8_locidx_t ltreeid, t8_locid
   T8_ASSERT (0 <= ltreeid && ltreeid < t8_forest_get_num_local_trees (forest));
 
   tree = t8_forest_get_tree (forest, ltreeid);
-  const t8_element_t * element = t8_forest_get_tree_element (tree, leid_in_tree);
+  const t8_element_t *element = t8_forest_get_tree_element (tree, leid_in_tree);
   T8_ASSERT (t8_forest_element_is_leaf (forest, element, ltreeid));
   return element;
 }
