+++ src/t8_forest/t8_forest.c
@@ -1406,7 +1406,7 @@ t8_forest_write_vtk (t8_forest_t forest, const char *fileprefix)
 
 t8_forest_t
 t8_forest_new_uniform (t8_cmesh_t cmesh, t8_scheme_cxx_t *scheme, const int level, const int do_face_ghost,
-                       sc_MPI_Comm comm)
+                       const int multilevel, sc_MPI_Comm comm)
 {
   t8_forest_t forest;
 
