+++ tutorials/general/t8_step5_element_data_c_interface.c
@@ -75,7 +75,7 @@ t8_step5_build_forest (sc_MPI_Comm comm, int level)
     0.4              /* Coarsen if outside this radius. */
   };
   /* Start with a uniform forest. */
-  t8_forest_t forest = t8_forest_new_uniform (cmesh, scheme, level, 0, comm);
+  t8_forest_t forest = t8_forest_new_uniform (cmesh, scheme, level, 0, 0, comm);
   t8_forest_t forest_apbg;
 
   /* Adapt, partition, balance and create ghost elements all in the same step. */
