### 2024-06-19 1eef4f8
Add apt-get update and upgrade to update_documentation action
