### 2024-06-19 5919150
add fetch-depth 0 to checkout action
