+++ .github/workflows/update_documentation.yml
@@ -45,6 +45,8 @@ jobs:
 # Setup and bootstrap
 #
     - uses: actions/checkout@v4
+       with: 
+        fetch-depth: 0
     - name: Update packages
       run: apt-get update && apt-get upgrade -y
     - name: Install zlib
