### 2024-06-19 a2ebe23
Update .github/workflows/update_documentation.yml
