+++ .github/workflows/update_documentation.yml
@@ -45,7 +45,7 @@ jobs:
 # Setup and bootstrap
 #
     - uses: actions/checkout@v4
-       with: 
+      with: 
         fetch-depth: 0
     - name: Update packages
       run: apt-get update && apt-get upgrade -y
