### 2024-06-19 8be6967
add sudo rights to workflow apt update
