+++ .github/workflows/update_documentation.yml
@@ -48,7 +48,7 @@ jobs:
       with: 
         fetch-depth: 0
     - name: Update packages
-      run: apt-get update && apt-get upgrade -y
+      run: sudo apt-get update && sudo apt-get upgrade -y
     - name: Install zlib
       run: sudo apt-get install libz-dev
     - name: Install doxygen
