### 2024-06-19 e713512
Add sudo apt-get update && sudo apt-get upgrade -y to all actions
Also remove fetch-depth from checkout action.
The default value 1 fetches the latest commit which is enough for
all our purposes.
