+++ .github/workflows/add_release_documentation.yml
@@ -56,6 +56,10 @@ jobs:
 #
 # Setup and bootstrap
 #
+
+    - name: Update packages
+      run: sudo apt-get update && sudo apt-get upgrade -y
+    # This step is necessary to get the newest package data
     - name: checkout
       if: ${{ env.MINOR_RELEASE == 'true' }}
       uses: actions/checkout@v4
@@ -114,7 +118,6 @@ jobs:
       uses: actions/checkout@v4
       with:
         ref: ${{ github.head_ref }}
-        fetch-depth: 0
         repository: DLR-AMR/t8code-website
         path: t8code-website
         token: ${{ secrets.T8DDY_TOKEN }}
