+++ .github/workflows/check_indentation.yml
@@ -52,6 +52,9 @@ jobs:
     timeout-minutes: 10
     steps:
     - uses: actions/checkout@v4
+    - name: Update packages
+      run: sudo apt-get update && sudo apt-get upgrade -y
+    # This step is necessary to get the newest package data
     - name: Install indent
       run: pip install clang-format==17.0.1
     - name: Indentation check
