+++ .github/workflows/test_tarball.yml
@@ -52,11 +52,9 @@ jobs:
 # Setup and bootstrap
 #
     - uses: actions/checkout@v4
-      with: 
-        fetch-depth: 0
     - name: Update packages
-      run: apt-get update && apt-get upgrade -y
-      # This seems to be necessary because of the docker container
+      run: sudo apt-get update && sudo apt-get upgrade -y
+    # This step is necessary to get the newest package data
     - name: disable ownership checks
       run: git config --global --add safe.directory '*'
     - name: init submodules
@@ -106,8 +104,8 @@ jobs:
     timeout-minutes: 90
     steps: 
     - name: Update packages
-      run: apt-get update && apt-get upgrade -y
-    # This seems to be necessary because of the docker container
+      run: sudo apt-get update && sudo apt-get upgrade -y
+    # This step is necessary to get the newest package data
     - name: Download tarball
       uses: actions/download-artifact@v4
       with:
