### 2024-06-19 b1af798
change to ubuntu-22.04 in parallel test
