### 2024-06-19 6ee5409
install sudo in parallel test
