+++ .github/workflows/tests_t8code_parallel.yml
@@ -61,6 +61,8 @@ jobs:
 #
     - uses: actions/checkout@v4
     - name: Update packages
+      run: apt update && apt install sudo
+    - name: Update packages
       run: sudo apt-get update && sudo apt-get upgrade -y
     # This step is necessary to get the newest package data
     - name: disable ownership checks
