### 2024-06-19 3027434
use fetch_tags option
