### 2024-06-19 9f28186
use fetch tags option to fetch the repo tags and create the version number correctly
