+++ .github/workflows/ci_playground.yml
@@ -77,6 +77,10 @@ jobs:
     runs-on: ubuntu-20.04
     steps:
       - uses: actions/checkout@v4
+      - name: install sudo
+        run: apt update && apt install sudo
+      # On the github Ubuntu 20.04, sudo is not available by default
+      # we need it, however, to update/upgrade our packages.      
       - name: Update packages
         run: sudo apt-get update && sudo apt-get upgrade -y
       # This step is necessary to get the newest package data
