+++ .github/workflows/update_documentation.yml
@@ -44,7 +44,7 @@ jobs:
 #
 # Setup and bootstrap
 #
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v4    
     - name: Update packages
       run: sudo apt-get update && sudo apt-get upgrade -y
     # This step is necessary to get the newest package data
