### 2024-06-19 a1d6ad8
fetch tags in parallel release
