+++ .github/workflows/tests_t8code_linkage_parallel_release.yml
@@ -62,6 +62,7 @@ jobs:
     - uses: actions/checkout@v4
       with:
           fetch-tags: true
+          fetch-deps: 0
     - name: install sudo
       run: apt update && apt install sudo
     # On the github Ubuntu 20.04, sudo is not available by default
