### 2024-06-19 c3f0632
reactive fetch-deps for all workflows
