+++ .github/workflows/tests_t8code_serial.yml
@@ -60,6 +60,9 @@ jobs:
 # Setup and bootstrap
 #
     - uses: actions/checkout@v4
+      with:
+        fetch-tags: true # required to get version tags
+        fetch-depth: 0    # required to get all history, especially the version tags
     - name: install sudo
       run: apt update && apt install sudo
     # On the github Ubuntu 20.04, sudo is not available by default
