+++ .github/workflows/update_documentation.yml
@@ -45,8 +45,9 @@ jobs:
 # Setup and bootstrap
 #
     - uses: actions/checkout@v4    
-      with: 
-       fetch-tags: true
+      with:
+        fetch-tags: true # required to get version tags
+        fetch-depth: 0    # required to get all history, especially the version tags
     - name: Update packages
       run: sudo apt-get update && sudo apt-get upgrade -y
     # This step is necessary to get the newest package data
