+++ .github/workflows/add_release_documentation.yml
@@ -66,7 +66,7 @@ jobs:
       with:
         path: t8code
         fetch-tags: true
-
+        fetch-depth: 0
     - name: Install dependencies
       if: ${{ env.MINOR_RELEASE == 'true' }}
       run: sudo apt-get install libz-dev doxygen-latex
