+++ .github/workflows/tests_t8code_linkage_parallel_release.yml
@@ -61,8 +61,8 @@ jobs:
 #
     - uses: actions/checkout@v4
       with:
-          fetch-tags: true
-          fetch-deps: 0
+          fetch-tags: true # required to get version tags
+          fetch-depth: 0    # required to get all history, especially the version tags
     - name: install sudo
       run: apt update && apt install sudo
     # On the github Ubuntu 20.04, sudo is not available by default
