+++ .github/workflows/check_indentation.yml
@@ -55,8 +55,9 @@ jobs:
       with:
         fetch-tags: true # required to get version tags
         fetch-depth: 0    # required to get all history, especially the version tags      
-    - name: install sudo
-      run: apt update && apt install sudo
+    # Deactivated the install sudo because it failed.
+    #- name: install sudo
+    #  run: apt update && apt install sudo
     # On the github Ubuntu 20.04, sudo is not available by default
     # we need it, however, to update/upgrade our packages.
     - name: Update packages
