### 2024-06-20 e45f1db
Fix calculation of attribute_data_offset
