### 2024-06-20 3371d37
Add documentation and check that t8_cmesh_set_partition_uniform is only called when committing a derived cmesh.
