+++ src/t8_cmesh.h
@@ -205,10 +205,10 @@ t8_cmesh_set_partition_range (t8_cmesh_t cmesh, int set_face_knowledge, t8_gloid
 void
 t8_cmesh_set_partition_offsets (t8_cmesh_t cmesh, t8_shmem_array_t tree_offsets);
 
-/** Declare if the cmesh is understood as a partitioned cmesh where the partition
- * table is derived from an assumed uniform refinement of a given level.
+/** Declare if a derived cmesh should be partitioned according to a
+ * uniform refinement of a given level for the provided scheme.
  * This call is only valid when the cmesh is not yet committed via a call
- * to \ref t8_cmesh_commit.
+ * to \ref t8_cmesh_commit and when the cmesh will be derived.
  * \param [in,out] cmesh          The cmesh to be updated.
  * \param [in]     element_level  The refinement_level.
  * \param [in]     ts             The element scheme describing the refinement pattern.
