### 2024-06-21 b47ed06
split hypercube_pad tests into two parts
therefore instead of multiplying by 8 it roughly adds 8 tests
