+++ .github/workflows/tests_t8code_linkage_parallel_debug.yml
@@ -54,7 +54,7 @@ jobs:
     if: (github.event_name == 'schedule' && github.repository == 'DLR-AMR/t8code') || (github.event_name != 'schedule')
     runs-on: ubuntu-20.04
     container: dlramr/t8code-ubuntu:t8-dependencies
-    timeout-minutes: 120
+    timeout-minutes: 90
     steps:
 #
 # Setup and bootstrap
