### 2024-06-21 #6887 @danielskatz
This problem also comes up in reproducibility and artifact evaluation. One option is simply to verify what can be verified (single system performance, scaling within that system) and to say that this is all that can be verified. Perhaps a reviewer might have access to a small cluster to run some small scaling experiments provided by the author? And/or perhaps the author can provide logs and outputs of large-scale runs?
