### 2024-06-24 d0a95a8
first ghost-interface draft
