### 2024-06-25 DLR-AMR/t8code#1101 @DamynChipman
Doxygen requirement during CMake build
I am attempting to install `t8code` on a Linux based cluster using CMake. I was able to install via `bootstrap; configure; make`, however I want to use the CMake interface. During CMake configuration, I get the following error:

```
CMake Error at /lfs/software/opt/linux-rocky8-x86_64/gcc-12.1.0/cmake-3.24.2-7bf344tdbrmmar6ktnedhjreayruw2l3/share/cmake-3.24/Modules/Find
PackageHandleStandardArgs.cmake:230 (message):
  Could NOT find Doxygen (missing: DOXYGEN_EXECUTABLE)
Call Stack (most recent call first):
  /lfs/software/opt/linux-rocky8-x86_64/gcc-12.1.0/cmake-3.24.2-7bf344tdbrmmar6ktnedhjreayruw2l3/share/cmake-3.24/Modules/FindPackageHandle
StandardArgs.cmake:594 (_FPHSA_FAILURE_MESSAGE)
  /lfs/software/opt/linux-rocky8-x86_64/gcc-12.1.0/cmake-3.24.2-7bf344tdbrmmar6ktnedhjreayruw2l3/share/cmake-3.24/Modules/FindDoxygen.cmake
:646 (find_package_handle_standard_args)
  doc/CMakeLists.txt:2 (find_package)
```

Here is my CMake configuration call:

```
cmake -B build -S . -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=local -DCMAKE_C_COMPILER=mpicc -DCMAKE_CXX_COMPILER=mpicxx
```

Is there an option to turn off Doxygen, or is it required?

I am reviewing `t8code` for JOSS: https://github.com/openjournals/joss-reviews/issues/6887
