### 2024-06-26 DLR-AMR/t8code#1101 @jmark
Hello and welcome to t8code!

There is a `BUILD_DOCUMENTATION` option in t8code's CMake build system. If you set it to `off` Doxygen is not required.

Let me know if you have further questions!
