### 2024-06-26 #6887 @jmark
> @danielskatz, reviewers have raised the point that it is not possible to verify the large-scale performance claims of the software without access to a large computing cluster. Do you have any guidance on how this has been handled for other JOSS submissions in the past?
>
> Alternatively, @jmark, do you have any input on how our reviewers might test the scalability of `t8code` without access to exascale compute resources?
>
> Thanks all!

Unfortunately, not really. Only some benchmarks on a cluster will show the scalability or not. I can offer to generate up-to-date benchmark data on clusters I have access to with setup parameters given by the reviewers.
