### 2024-06-26 94dcde3
Actually memcpy the data, and not the container
