### 2024-06-26 DLR-AMR/t8code#1101 @jmark
> Maybe we should rename that option to `T8CODE_BUILD_DOCUMENTATION` to make things clearer and ease interop with downstream software. What do you think @jmark ?

Good idea!
