### 2024-06-26 70703c6
Use correct wrapped shmem_array functionality to copy data
