### 2024-06-26 2f241a0
Also include ghost attributes in the byte allocation calculation
