+++ src/t8_cmesh/t8_cmesh_trees.c
@@ -419,13 +419,15 @@ t8_cmesh_trees_get_part_alloc (t8_cmesh_trees_t trees, t8_part_tree_t part)
   byte_alloc = part->num_trees * sizeof (t8_ctree_struct_t) + part->num_ghosts * sizeof (t8_cghost_struct_t);
   for (ltree = 0; ltree < part->num_trees; ltree++) {
     tree = t8_cmesh_trees_get_tree (trees, ltree + part->first_tree_id);
+    byte_alloc += t8_cmesh_trees_neighbor_bytes (tree);
     byte_alloc += t8_cmesh_trees_attribute_size (tree);
     byte_alloc += tree->num_attributes * sizeof (t8_attribute_info_struct_t);
-    byte_alloc += t8_cmesh_trees_neighbor_bytes (tree);
   }
   for (lghost = 0; lghost < part->num_ghosts; lghost++) {
     ghost = t8_cmesh_trees_get_ghost (trees, lghost + part->first_ghost_id);
     byte_alloc += t8_cmesh_trees_gneighbor_bytes (ghost);
+    byte_alloc += t8_cmesh_trees_ghost_attribute_size (ghost);
+    byte_alloc += ghost->num_attributes * sizeof (t8_attribute_info_struct_t);
   }
   return byte_alloc;
 }
