### 2024-06-26 #6887 @danielskatz
Thanks for that offer @jmark. Also, given that @DamynChipman can run on a small cluster, this seems like it may be sufficient.  @elbeejay, what do you think?
