### 2024-06-27 DLR-AMR/t8code#1101 @DamynChipman
Using `-DBUILD_DOCUMENTATION=off` worked, thanks!
