### 2024-06-28 0578268
Revert "enable partitioned cmeshs in tests"
