### 2024-06-28 4b1ff0a
remove dquad implementation
