+++ src/t8_schemes/t8_default/t8_default_quad/t8_dquad_bits.h
@@ -43,7 +43,7 @@ T8_EXTERN_C_BEGIN ();
  *                         of the points on the quad.
  */
 void
-t8_dquad_compute_reference_coords (const t8_dquad_t *elem, const double *ref_coords, const size_t num_coords,
+t8_dquad_compute_reference_coords (const p4est_quadrant_t *elem, const double *ref_coords, const size_t num_coords,
                                    double *out_coords);
 
 T8_EXTERN_C_END ();
