+++ src/t8_schemes/t8_default/t8_default_quad/t8_dquad_bits.c
@@ -23,6 +23,17 @@
 #include <p4est_bits.h>
 #include <t8_schemes/t8_default/t8_default_quad/t8_dquad_bits.h>
 #include <p4est_bits.h>
+#include <t8_schemes/t8_default/t8_default_quad/t8_default_quad_cxx.hxx>
+
+static void
+t8_element_copy_surround (const p4est_quadrant_t *q, p4est_quadrant_t *r)
+{
+  T8_QUAD_SET_TDIM (r, T8_QUAD_GET_TDIM (q));
+  if (T8_QUAD_GET_TDIM (q) == 3) {
+    T8_QUAD_SET_TNORMAL (r, T8_QUAD_GET_TNORMAL (q));
+    T8_QUAD_SET_TCOORD (r, T8_QUAD_GET_TCOORD (q));
+  }
+}
 
 void
 t8_dquad_compute_reference_coords (const p4est_quadrant_t *elem, const double *ref_coords, const size_t num_coords,
@@ -38,5 +49,93 @@ t8_dquad_compute_reference_coords (const p4est_quadrant_t *elem, const double *r
 
     out_coords[offset_2d + 0] /= (double) P4EST_ROOT_LEN;
     out_coords[offset_2d + 1] /= (double) P4EST_ROOT_LEN;
+  }
+}
+
+void
+t8_dquad_child (const p4est_quadrant_t *elem, int childid, p4est_quadrant_t *child)
+{
+  const p4est_qcoord_t shift = P4EST_QUADRANT_LEN (elem->level + 1);
+
+  T8_ASSERT (p4est_quadrant_is_extended (elem));
+  T8_ASSERT (elem->level < P4EST_QMAXLEVEL);
+  T8_ASSERT (childid >= 0 && childid < P4EST_CHILDREN);
+
+  child->x = childid & 0x01 ? (elem->x | shift) : elem->x;
+  child->y = childid & 0x02 ? (elem->y | shift) : elem->y;
+  child->level = elem->level + 1;
+
+  if (elem != child) {
+    T8_ASSERT (p4est_quadrant_is_parent (elem, child));
+  }
+  t8_element_copy_surround (elem, child);
+}
+
+void
+t8_dquad_parent (const p4est_quadrant_t *elem, p4est_quadrant_t *parent)
+{
+  T8_ASSERT (p4est_quadrant_is_extended (elem));
+  p4est_quadrant_parent (elem, parent);
+  t8_element_copy_surround (elem, parent);
+}
+
+void
+t8_dquad_sibling (const p4est_quadrant_t *elem, int sibid, p4est_quadrant_t *sibling)
+{
+  T8_ASSERT (p4est_quadrant_is_extended (elem));
+  p4est_quadrant_sibling (elem, sibling, sibid);
+  t8_element_copy_surround (elem, sibling);
+}
+
+void
+t8_dquad_copy (const p4est_quadrant_t *source, p4est_quadrant_t *dest)
+{
+  T8_ASSERT (p4est_quadrant_is_extended (source));
+  if (source == dest) {
+    /* Do nothing if they are already the same quadrant. */
+    return;
+  }
+  *dest = *source;
+  t8_element_copy_surround (source, dest);
+}
+
+void
+t8_dquad_successor (const p4est_quadrant_t *elem, p4est_quadrant_t *succ, const int level, const int multilevel)
+{
+  T8_ASSERT (p4est_quadrant_is_extended (elem));
+  T8_ASSERT (0 <= elem->level && succ->level <= P4EST_QMAXLEVEL);
+  T8_ASSERT (0 <= level && level <= P4EST_QMAXLEVEL);
+  T8_ASSERT (elem->level <= level);
[…]
