+++ tutorials/CMakeLists.txt
@@ -22,3 +22,4 @@ add_t8_tutorial( NAME t8_step7_interpolation            SOURCES general/t8_step7
 add_t8_tutorial( NAME t8_tutorial_build_cmesh           SOURCES general/t8_tutorial_build_cmesh.cxx general/t8_tutorial_build_cmesh_main.cxx)
 add_t8_tutorial( NAME t8_tutorial_search                SOURCES general/t8_tutorial_search.cxx general/t8_step3_adapt_forest.cxx )
 add_t8_tutorial( NAME t8_features_curved_meshes         SOURCES features/t8_features_curved_meshes.cxx)
+add_t8_tutorial( NAME t8_features_multilevel            SOURCES features/t8_features_multilevel.cxx)
