### 2024-07-01 b8c6e37
update ghost_interface_unref, update set_ghost (not final), add debuging prints
