### 2024-07-01 45ef33f
bugfix: syntax in cmake ci
