### 2024-07-01 1907f4d
bugfix: syntax in cmake ci
