+++ doc/dev/search/all_17.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['z_1279',['z',['../structt8__dhex.html#adbb4779dab14e8c25bd8d98fcc182805',1,'t8_dhex::z()'],['../structt8__dtet.html#a4e43487fec279f5e0f6065446ff8ecb5',1,'t8_dtet::z()']]]
+];

[…]
