### 2024-07-03 f0bb93c
Added project version to CMake build system
