### 2024-07-03 c6124b6
Added libt8 shared library version and soname in CMake build system
