### 2024-07-03 c8e4dc4
Remove duplicate code
