+++ src/t8_forest/t8_forest_ghost.h
@@ -162,10 +162,38 @@ t8_forest_ghost_unref (t8_forest_ghost_t *pghost);
 void
 t8_forest_ghost_destroy (t8_forest_ghost_t *pghost);
 
+/** Part of step 2 of the ghost_creat_ext 
+ * for ghost_type face
+ * Is declared, so that ghost_interface_face can use ist
+ * \see t8_forest_ghost_interface_faces::t8_ghost_step_2
+ * \param [in,out]    forest     The forest.
+ */
+void
+t8_forest_ghost_fill_remote_v3 (t8_forest_t forest);
+
+/** Part of step 2 of the ghost_creat_ext 
+ * for ghost_type face
+ * Is declared, so that ghost_interface_face can use ist
+ * \see t8_forest_ghost_interface_faces::t8_ghost_step_2
+ * \param [in,out]    forest     The forest.
+ */
+void
+t8_forest_ghost_fill_remote (t8_forest_t forest, t8_forest_ghost_t ghost, int ghost_method);
+
+
 /** Create one layer of ghost elements for a forest.
  * \see t8_forest_set_ghost
  * \param [in,out]    forest     The forest.
  * \a forest must be committed before calling this function.
+ */
+void
+t8_forest_ghost_create_ext (t8_forest_t forest);
+
+/** Create one layer of ghost elements for a forest.
+ * \see t8_forest_set_ghost
+ * \param [in,out]    forest     The forest.
+ * \a forest must be committed before calling this function.
+ * \a forest->ghost_interface must have the \a type FACE and the \a version 2
  */
 void
 t8_forest_ghost_create (t8_forest_t forest);
@@ -176,12 +204,18 @@ t8_forest_ghost_create (t8_forest_t forest);
  *                Mesh Refinement On Forests of Octrees
  * \param [in,out]    forest     The balanced forest/
  * \a forest must be committed before calling this function.
+ * \a forest->ghost_interface must have the \a type FACE and the \a version 1
  * \note The user should prefer \ref t8_forest_ghost_create even for balanced forests.
  */
 void
 t8_forest_ghost_create_balanced_only (t8_forest_t forest);
 
-/* experimental version using the ghost_v3 algorithm */
+/** Creating one layer of ghost elements for a forest. 
+ * experimental version using the ghost_v3 algorithm
+ * \param [in,out]    forest     The forest.
+ * \a forest must be committed before calling this function.
+ * \a forest->ghost_interface must have the \a type FACE and the \a version 1
+ */
 void
 t8_forest_ghost_create_topdown (t8_forest_t forest);
 
