### 2024-07-04 65fc4e4
delete unnecesssary prints
