### 2024-07-04 3774154
bugfix: syntax in cmake ci
