### 2024-07-04 c61a5a4
commed out the members ghost_type and ghost_algorithm in forest
