### 2024-07-04 d169b86
bugfix: syntax in cmake ci
