### 2024-07-04 DLR-AMR/t8code#1109 @holke
Hi @DamynChipman,
are you by any chance using an M1 or M2 Mac processor? Possibly with OpenMPI?
We recently noticed that this combination seems to have issues with the MPI shared memory implementation.

Shared memory not being active does not result in accuracy loss.
It will increase the memory usage. However, the real use of shared memory for us kicks in when running on a cluster on > 1000 CPUs.
