### 2024-07-05 e42bae2
rename t8_forest_cxx.cxx to t8_forest.cxx and merge t8_forest.c into it
