+++ src/CMakeLists.txt
@@ -118,10 +118,9 @@ target_sources( T8 PRIVATE
     t8_cmesh/t8_cmesh_readmshfile.cxx 
     t8_data/t8_shmem.c 
     t8_data/t8_containers.cxx 
-    t8_forest/t8_forest.c 
     t8_forest/t8_forest_adapt.cxx 
     t8_forest/t8_forest_partition.cxx 
-    t8_forest/t8_forest_cxx.cxx 
+    t8_forest/t8_forest.cxx 
     t8_forest/t8_forest_private.c 
     t8_forest/t8_forest_vtk.cxx 
     t8_forest/t8_forest_ghost.cxx 
