+++ src/CMakeLists.txt
@@ -138,26 +138,26 @@ target_sources( T8 PRIVATE
     t8_geometry/t8_geometry_implementations/t8_geometry_lagrange.cxx 
     t8_geometry/t8_geometry_implementations/t8_geometry_zero.cxx 
     t8_geometry/t8_geometry_implementations/t8_geometry_examples.cxx 
-    t8_schemes/t8_default/t8_default_cxx.cxx
-    t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.cxx
-    t8_schemes/t8_default/t8_default_hex/t8_default_hex_cxx.cxx
+    t8_schemes/t8_default/t8_default.cxx
+    t8_schemes/t8_default/t8_default_common/t8_default_common.cxx
+    t8_schemes/t8_default/t8_default_hex/t8_default_hex.cxx
     t8_schemes/t8_default/t8_default_hex/t8_dhex_bits.c
-    t8_schemes/t8_default/t8_default_line/t8_default_line_cxx.cxx
+    t8_schemes/t8_default/t8_default_line/t8_default_line.cxx
     t8_schemes/t8_default/t8_default_line/t8_dline_bits.c
-    t8_schemes/t8_default/t8_default_prism/t8_default_prism_cxx.cxx
+    t8_schemes/t8_default/t8_default_prism/t8_default_prism.cxx
     t8_schemes/t8_default/t8_default_prism/t8_dprism_bits.c
-    t8_schemes/t8_default/t8_default_pyramid/t8_default_pyramid_cxx.cxx
+    t8_schemes/t8_default/t8_default_pyramid/t8_default_pyramid.cxx
     t8_schemes/t8_default/t8_default_pyramid/t8_dpyramid_bits.c
     t8_schemes/t8_default/t8_default_pyramid/t8_dpyramid_connectivity.c 
-    t8_schemes/t8_default/t8_default_quad/t8_default_quad_cxx.cxx
+    t8_schemes/t8_default/t8_default_quad/t8_default_quad.cxx
     t8_schemes/t8_default/t8_default_quad/t8_dquad_bits.c
-    t8_schemes/t8_default/t8_default_tet/t8_default_tet_cxx.cxx
+    t8_schemes/t8_default/t8_default_tet/t8_default_tet.cxx
     t8_schemes/t8_default/t8_default_tet/t8_dtet_bits.c
     t8_schemes/t8_default/t8_default_tet/t8_dtet_connectivity.c
-    t8_schemes/t8_default/t8_default_tri/t8_default_tri_cxx.cxx
+    t8_schemes/t8_default/t8_default_tri/t8_default_tri.cxx
     t8_schemes/t8_default/t8_default_tri/t8_dtri_bits.c
     t8_schemes/t8_default/t8_default_tri/t8_dtri_connectivity.c
-    t8_schemes/t8_default/t8_default_vertex/t8_default_vertex_cxx.cxx
+    t8_schemes/t8_default/t8_default_vertex/t8_default_vertex.cxx
     t8_schemes/t8_default/t8_default_vertex/t8_dvertex_bits.c
     t8_vtk/t8_vtk_polydata.cxx 
     t8_vtk/t8_vtk_unstructured.cxx 
