### 2024-07-05 #6887 @DamynChipman
> 👋 @DamynChipman, please update us on how your review is going (this is an automated reminder).

I have `t8code` installed on my laptop and a cluster and have gone through the tutorials. Good documentation and walkthroughs so far! I raised a couple of issues that the `t8code` team addressed.

@jmark, do you have any suggestions for benchmarks to run to reproduce some scaling analysis? I noticed there's a `benchmarks` directory in the source code, as well as some examples I have ran.
