### 2024-07-06 #6887 @alecontri
> 👋 @alecontri, please update us on how your review is going (this is an automated reminder).

I have `t8code` installed on my laptop. I raised an installation issue that the `t8code` team addressed, so next week I will go on and test the tutorials.

@jmark regarding the comments to the paper, do we have to open an issue on the `t8code` repository?
