### 2024-07-08 4d9e6b0
Add docstring
