### 2024-07-08 #6887 @jmark
> > 👋 @alecontri, please update us on how your review is going (this is an automated reminder).
>
> I have `t8code` installed on my laptop. I raised an installation issue that the `t8code` team addressed, so next week I will go on and test the tutorials.
>
> @jmark regarding the comments to the paper, do we have to open an issue on the `t8code` repository?

@alecontri You can open issues on the `t8code` directory directly, if you prefer to do so. As an alternative, I invite you to directly
open issues within our t8code JOSS paper repository. Then we have a clean separation of concerns.

https://github.com/DLR-AMR/t8code_JOSS
