### 2024-07-09 b8f0a34
Simplified C interface for registering geometry handlers.
