### 2024-07-10 0afb894
Did some refactoring.
