+++ CMakeLists.txt
@@ -5,7 +5,7 @@ include(cmake/GitProjectVersion.cmake)
 project( T8CODE
   DESCRIPTION "Parallel algorithms and data structures for tree-based AMR with arbitrary element shapes."
   LANGUAGES C CXX
-  VERSION "${T8CODE_VERSION_MAJOR}.${T8CODE_VERSION_MINOR}" )
+  VERSION "${T8CODE_VERSION_MAJOR}.${T8CODE_VERSION_MINOR}.${T8CODE_VERSION_PATCH}" )
 include( CTest )
 
 option( T8CODE_BUILD_AS_SHARED_LIBRARY "Whether t8code should be built as a shared or a static library" ON )
