+++ src/CMakeLists.txt
@@ -195,7 +195,7 @@ INSTALL_DESTINATION cmake
 write_basic_package_version_file(
 ${CMAKE_CURRENT_BINARY_DIR}/CMakeFiles/${PROJECT_NAME}ConfigVersion.cmake
 COMPATIBILITY SameMajorVersion
-VERSION "${T8CODE_VERSION_MAJOR}.${T8CODE_VERSION_MINOR}"
+VERSION "${T8CODE_VERSION_MAJOR}.${T8CODE_VERSION_MINOR}.${T8CODE_VERSION_PATCH}"
 )
 
 install( EXPORT ${PROJECT_NAME}-targets
