### 2024-07-11 2f07e9f
Removed duplicate code
