### 2024-07-11 0e1fb36
Fixed T8_{C,CPP}FLAGS compile definitions
