+++ src/CMakeLists.txt
@@ -9,11 +9,14 @@ endif()
 add_library( T8CODE::T8 ALIAS T8 )
 
 if( CMAKE_BUILD_TYPE STREQUAL "Release" )
-    target_compile_options( T8 PRIVATE -O3 )
+  set (T8_CPPFLAGS "${CMAKE_CXX_FLAGS} ${CMAKE_CXX_FLAGS_RELEASE}")
+  set (T8_CFLAGS "${CMAKE_C_FLAGS} ${CMAKE_C_FLAGS_RELEASE}")
 endif()
 
 if( CMAKE_BUILD_TYPE STREQUAL "Debug" )
-    target_compile_options( T8 PUBLIC -g -O0 )
+  set (T8_CPPFLAGS "${CMAKE_CXX_FLAGS} ${CMAKE_CXX_FLAGS_DEBUG}")
+  set (T8_CFLAGS "${CMAKE_C_FLAGS} ${CMAKE_C_FLAGS_DEBUG}")
+
     target_compile_definitions( T8 PUBLIC T8_ENABLE_DEBUG )
 endif()
 
@@ -21,9 +24,9 @@ set_target_properties( T8 PROPERTIES OUTPUT_NAME t8 )
 
 target_compile_definitions( T8 PUBLIC T8_CMAKE_BUILD )
 target_compile_definitions( T8 PUBLIC T8_CC="${CMAKE_C_COMPILER}" )
-target_compile_definitions( T8 PUBLIC T8_CFLAGS="${CMAKE_C_FLAGS}" )
+target_compile_definitions( T8 PUBLIC T8_CFLAGS="${T8_CFLAGS}" )
 target_compile_definitions( T8 PUBLIC T8_CPP="${CMAKE_CXX_COMPILER}" )
-target_compile_definitions( T8 PUBLIC T8_CPPFLAGS="${CMAKE_CXX_FLAGS}" )
+target_compile_definitions( T8 PUBLIC T8_CPPFLAGS="${T8_CPPFLAGS}" )
 target_compile_definitions( T8 PUBLIC T8_LDFLAGS="${CMAKE_SHARED_LINKER_FLAGS}" )
 target_compile_definitions( T8 PUBLIC T8_LIBS="Not available with CMake builds" )
 
