+++ src/CMakeLists.txt
@@ -13,6 +13,11 @@ if( CMAKE_BUILD_TYPE STREQUAL "Release" )
   set (T8_CFLAGS "${CMAKE_C_FLAGS} ${CMAKE_C_FLAGS_RELEASE}")
 endif()
 
+if( CMAKE_BUILD_TYPE STREQUAL "RelWithDebInfo" )
+  set (T8_CPPFLAGS "${CMAKE_CXX_FLAGS} ${CMAKE_CXX_FLAGS_RELWITHDEBINFO}")
+  set (T8_CFLAGS "${CMAKE_C_FLAGS} ${CMAKE_C_FLAGS_RELWITHDEBINFO}")
+endif()
+
 if( CMAKE_BUILD_TYPE STREQUAL "Debug" )
   set (T8_CPPFLAGS "${CMAKE_CXX_FLAGS} ${CMAKE_CXX_FLAGS_DEBUG}")
   set (T8_CFLAGS "${CMAKE_C_FLAGS} ${CMAKE_C_FLAGS_DEBUG}")
