### 2024-07-11 a187f07
Changed T8_PACKAGE_STRING compile definition
