### 2024-07-11 3c254f0
fix cast errors
