### 2024-07-12 ecadb64
Fixed T8_LIBS compile definition in CMake build system
