+++ src/CMakeLists.txt
@@ -27,21 +27,6 @@ endif()
 
 set_target_properties( T8 PROPERTIES OUTPUT_NAME t8 )
 
-target_compile_definitions( T8 PUBLIC T8_CMAKE_BUILD )
-target_compile_definitions( T8 PUBLIC T8_CC="${CMAKE_C_COMPILER}" )
-target_compile_definitions( T8 PUBLIC T8_CFLAGS="${T8_CFLAGS}" )
-target_compile_definitions( T8 PUBLIC T8_CPP="${CMAKE_CXX_COMPILER}" )
-target_compile_definitions( T8 PUBLIC T8_CPPFLAGS="${T8_CPPFLAGS}" )
-target_compile_definitions( T8 PUBLIC T8_LDFLAGS="${CMAKE_SHARED_LINKER_FLAGS}" )
-target_compile_definitions( T8 PUBLIC T8_LIBS="Not available with CMake builds" )
-
-target_compile_definitions( T8 PUBLIC T8_PACKAGE_STRING="t8 ${T8CODE_VERSION_RAW}" )
-target_compile_definitions( T8 PUBLIC T8_VERSION="${T8CODE_VERSION}" )
-target_compile_definitions( T8 PUBLIC T8_VERSION_MAJOR=${T8CODE_VERSION_MAJOR} )
-target_compile_definitions( T8 PUBLIC T8_VERSION_MINOR=${T8CODE_VERSION_MINOR} )
-target_compile_definitions( T8 PUBLIC T8_VERSION_PATCH=${T8CODE_VERSION_PATCH} )
-
-
 target_include_directories( T8 PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>
   $<INSTALL_INTERFACE:${CMAKE_INSTALL_PREFIX}/include>
@@ -145,6 +130,23 @@ target_sources( T8 PRIVATE
     t8_vtk/t8_vtk_parallel.cxx 
     t8_vtk/t8_vtk_reader.cxx 
 )
+
+target_compile_definitions( T8 PUBLIC T8_CMAKE_BUILD )
+target_compile_definitions( T8 PUBLIC T8_CC="${CMAKE_C_COMPILER}" )
+target_compile_definitions( T8 PUBLIC T8_CFLAGS="${T8_CFLAGS}" )
+target_compile_definitions( T8 PUBLIC T8_CPP="${CMAKE_CXX_COMPILER}" )
+target_compile_definitions( T8 PUBLIC T8_CPPFLAGS="${T8_CPPFLAGS}" )
+target_compile_definitions( T8 PUBLIC T8_LDFLAGS="${CMAKE_SHARED_LINKER_FLAGS}" )
+
+target_compile_definitions( T8 PUBLIC T8_PACKAGE_STRING="t8 ${T8CODE_VERSION_RAW}" )
+target_compile_definitions( T8 PUBLIC T8_VERSION="${T8CODE_VERSION}" )
+target_compile_definitions( T8 PUBLIC T8_VERSION_MAJOR=${T8CODE_VERSION_MAJOR} )
+target_compile_definitions( T8 PUBLIC T8_VERSION_MINOR=${T8CODE_VERSION_MINOR} )
+target_compile_definitions( T8 PUBLIC T8_VERSION_PATCH=${T8CODE_VERSION_PATCH} )
+
+get_target_property( T8_LIBS_LIST T8 LINK_LIBRARIES )
+string( REPLACE ";" " " T8_LIBS "${T8_LIBS_LIST}" )
+target_compile_definitions( T8 PUBLIC T8_LIBS="${T8_LIBS}" )
 
 install( FILES
     t8.h
