### 2024-07-12 119c5a4
Added new convenience example.
