### 2024-07-15 327aab3
Use set_join_by_stash instead of set_join_by_vertices
