### 2024-07-15 9b6649e
add serial cmake test
