### 2024-07-15 bf92278
remove faulty line in serial cmake ci
