### 2024-07-15 64b81f4
change faulty path in serial cmake ci
