### 2024-07-15 51cfd51
fixes in cmake ci
