+++ .github/workflows/tests_t8code_cmake_parallel.yml
@@ -95,7 +95,7 @@ jobs:
           ${{ env.SC_DEBUG }}
           ${{ env.SC_RELEASE }}
         # You can increase the counter at the end to force a new key and hence recomputing the cache
-        key: sc-${{ env.sc_commit }}-005
+        key: sc-cmake-${{ env.sc_commit }}-005
     - name: Set ignore cache variable
       # If this variable is set 1
       # (i.e. SC_IGNORE_CACHE=1) then the cache will be ignored and
@@ -108,7 +108,7 @@ jobs:
     - name: if ignore cache, delete folders
       if: ${{ env.SC_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist
-      run: rm -r $SC_DEBUG $SC_RELEASE $SC_SERIAL_DEBUG $SC_SERIAL_RELEASE || true
+      run: rm -r $SC_DEBUG $SC_RELEASE || true
     - name: make folders
       run: mkdir $SC_DEBUG && mkdir $SC_RELEASE
       if: ${{ steps.sc_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
@@ -145,7 +145,7 @@ jobs:
           ${{ env.P4EST_DEBUG }}
           ${{ env.P4EST_RELEASE }}
         # You can increase the counter at the end to force a new key and hence recomputing the cache
-        key: p4est-${{ env.p4est_commit }}-${{ env.sc_commit }}-005
+        key: p4est-cmake-${{ env.p4est_commit }}-${{ env.sc_commit }}-005
     - name: Set ignore cache variable
       # If this variable is set to 1 then the cache is ignored and p4est is installed regardless of cache.
       # (set with i.e. P4EST_IGNORE_CACHE=1)
@@ -161,7 +161,7 @@ jobs:
       if: ${{ env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
       run: rm -r $P4EST_DEBUG $P4EST_RELEASE || true 
-    - name: make folder
+    - name: make folders
       run: mkdir $P4EST_DEBUG $P4EST_RELEASE
     ## p4est debug
     - name: p4est cmake debug
