### 2024-07-15 275f409
fixes in cmake ci
