+++ .github/workflows/tests_t8code_cmake_parallel.yml
@@ -189,9 +189,9 @@ jobs:
            && echo LESS_TEST_OPTION="$LESS_TEST_OPTION" >> $GITHUB_ENV
     - name: build config variables
       #TODO: Werror
-      run: export CONFIG_OPTIONS="${LESS_TEST_OPTION}"
-           && export CONFIG_DEBUG="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Debug -DT8CODE_USE_SYSTEM_SC=ON -DSC_DIR=$SC_DEBUG/install/cmake -DT8CODE_USE_SYSTEM_P4EST=ON -DP4EST_DIR=$P4EST_DEBUG/install/cmake"
-           && export CONFIG_RELEASE="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Release -DT8CODE_USE_SYSTEM_SC=ON -DSC_DIR=$SC_RELEASE/install/cmake -DT8CODE_USE_SYSTEM_P4EST=ON -DP4EST_DIR=$P4EST_RELEASE/install/cmake"
+      run: export CONFIG_OPTIONS="${LESS_TEST_OPTION} -DT8CODE_USE_SYSTEM_SC=ON -DT8CODE_USE_SYSTEM_P4EST=ON"
+           && export CONFIG_DEBUG="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Debug -DSC_DIR=$SC_DEBUG/install/cmake -DP4EST_DIR=$P4EST_DEBUG/install/cmake"
+           && export CONFIG_RELEASE="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Release -DSC_DIR=$SC_RELEASE/install/cmake -DP4EST_DIR=$P4EST_RELEASE/install/cmake"
            && echo CONFIG_OPTIONS="$CONFIG_OPTIONS" >> $GITHUB_ENV
            && echo CONFIG_DEBUG="$CONFIG_DEBUG" >> $GITHUB_ENV
            && echo CONFIG_RELEASE="$CONFIG_RELEASE" >> $GITHUB_ENV
