### 2024-07-15 7535a86
add cmake parallel linkage ci
