+++ .github/workflows/tests_t8code_cmake_linkage_parallel_debug.yml
@@ -1,0 +1,266 @@
+name: t8code cmake tests linkage parallel debug
+
+
+#  This file is part of t8code.
+#  t8code is a C library to manage a collection (a forest) of multiple
+#  connected adaptive space-trees of general element types in parallel.
+#
+#  Copyright (C) 2024 the developers
+#
+#  t8code is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation; either version 2 of the License, or
+#  (at your option) any later version.
+#
+#  t8code is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with t8code; if not, write to the Free Software Foundation, Inc.,
+#  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+
+#
+# This github CI script installs t8code and runs its tests for various configurations.
+# We compile sc and p4est as thirdparty libraries and use caching to only trigger a
+# new installation of them when their versions have changed in t8code.
+#
+# Note: To manually enforce sc and p4est installation, either increase the counter
+#       in the "key:" entries of the sc and p4est steps or set the variables
+#       SC_IGNORE_CACHE and P4EST_IGNORE_CACHE to 1 in the respective steps.
+
+env:
+  MAKEFLAGS: "-j2"
+
+on:
+  push:
+    branches:
+      - main
+      - develop
+      - feature-*CI* # for testing this script, all feature branches with "CI" in their name
+  pull_request:
+    branches:
+      - main
+      - develop
+  workflow_dispatch: # Be able to trigger this manually on github.com
+  # Run every night at 1:10
+  schedule:
+      - cron:  '10 1 * * *'
+
+jobs:
+  build:
+
+    if: (github.event_name == 'schedule' && github.repository == 'DLR-AMR/t8code') || (github.event_name != 'schedule')
+    runs-on: ubuntu-20.04
+    container: dlramr/t8code-ubuntu:t8-dependencies
+    timeout-minutes: 90
+    steps:
+#
+# Setup and bootstrap
+#
+    - uses: actions/checkout@v4
+      with: 
+        fetch-depth: 0
+    - name: Update packages
+      run: apt-get update && apt-get upgrade -y
+      # This seems to be necessary because of the docker container
+    - name: disable ownership checks
+      run: git config --global --add safe.directory '*'
+    - name: init submodules
+      run: git submodule init
+    - name: update submodules
+      run: git submodule update
+    - name: alter p4est submodule remote
+      run: git submodule set-url p4est https://github.com/sandro-elsweijer/p4est.git
+    - name: Checkout p4est submodule
+      run: |
+        cd p4est
+        git fetch --all
+        git checkout aca63b37d2fc7b2aef601e320c30f92826248388
+        cd ..
+#
[…]
