### 2024-07-15 f0449c3
trickle down const effects to t8_cmesh_add_attributes
