+++ src/t8_cmesh/t8_cmesh_trees.c
@@ -64,7 +64,8 @@ t8_cmesh_trees_get_part (const t8_cmesh_trees_t trees, const int proc)
 }
 
 void
-t8_cmesh_trees_init (t8_cmesh_trees_t *ptrees, int num_procs, t8_locidx_t num_trees, t8_locidx_t num_ghosts)
+t8_cmesh_trees_init (t8_cmesh_trees_t *ptrees, const int num_procs, const t8_locidx_t num_trees,
+                     const t8_locidx_t num_ghosts)
 {
   t8_cmesh_trees_t trees;
 
@@ -85,7 +86,7 @@ t8_cmesh_trees_init (t8_cmesh_trees_t *ptrees, int num_procs, t8_locidx_t num_tr
 }
 
 void
-t8_cmesh_trees_add_tree (t8_cmesh_trees_t trees, t8_locidx_t ltree_id, int proc, t8_eclass_t eclass)
+t8_cmesh_trees_add_tree (t8_cmesh_trees_t trees, const t8_locidx_t ltree_id, const int proc, const t8_eclass_t eclass)
 {
   t8_part_tree_t part;
   t8_ctree_t tree;
@@ -106,8 +107,8 @@ t8_cmesh_trees_add_tree (t8_cmesh_trees_t trees, t8_locidx_t ltree_id, int proc,
 }
 
 void
-t8_cmesh_trees_add_ghost (t8_cmesh_trees_t trees, t8_locidx_t lghost_index, t8_gloidx_t gtree_id, int proc,
-                          t8_eclass_t eclass, t8_locidx_t num_local_trees)
+t8_cmesh_trees_add_ghost (const t8_cmesh_trees_t trees, const t8_locidx_t lghost_index, const t8_gloidx_t gtree_id,
+                          const int proc, const t8_eclass_t eclass, const t8_locidx_t num_local_trees)
 {
   t8_part_tree_t part;
   t8_cghost_t ghost;
@@ -182,8 +183,9 @@ t8_part_tree_get_ghost (t8_part_tree_t P, t8_locidx_t ghost_id)
 }
 
 void
-t8_cmesh_trees_start_part (t8_cmesh_trees_t trees, int proc, t8_locidx_t lfirst_tree, t8_locidx_t num_trees,
-                           t8_locidx_t lfirst_ghost, t8_locidx_t num_ghosts, int alloc)
+t8_cmesh_trees_start_part (const t8_cmesh_trees_t trees, const int proc, const t8_locidx_t lfirst_tree,
+                           const t8_locidx_t num_trees, const t8_locidx_t lfirst_ghost, const t8_locidx_t num_ghosts,
+                           const int alloc)
 {
   t8_part_tree_t part;
   T8_ASSERT (trees != NULL);
@@ -217,7 +219,7 @@ t8_cmesh_trees_start_part (t8_cmesh_trees_t trees, int proc, t8_locidx_t lfirst_
  * init_attributes, call finish_part, successively call add_attributes
  * and also set all face neighbors (TODO: write function)*/
 size_t
-t8_cmesh_trees_finish_part (t8_cmesh_trees_t trees, int proc)
+t8_cmesh_trees_finish_part (const t8_cmesh_trees_t trees, const int proc)
 {
   t8_part_tree_t part;
   t8_ctree_t tree;
@@ -326,7 +328,7 @@ t8_cmesh_trees_finish_part (t8_cmesh_trees_t trees, int proc)
 }
 
 void
-t8_cmesh_trees_set_all_boundary (t8_cmesh_t cmesh, t8_cmesh_trees_t trees)
[…]
