+++ .github/workflows/tests_t8code_cmake_serial.yml
@@ -106,7 +106,7 @@ jobs:
       if: ${{ steps.sc_cmake_serial_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       run: echo No cache found or cache will be ignored. SC_IGNORE_CACHE=$SC_IGNORE_CACHE
     - name: if ignore cache, delete folders
-      if: ${{ env.SC_IGNORE_CACHE == 1 }}
+      if: ${{ steps.sc_cmake_serial_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist
       run: rm -r $SC_SERIAL_DEBUG $SC_SERIAL_RELEASE || true
     - name: make folders
@@ -158,7 +158,7 @@ jobs:
       run: echo "Install p4est"
       if: ${{ steps.p4est_cmake_serial_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
     - name: if ignore cache, delete folders
-      if: ${{ env.P4EST_IGNORE_CACHE == 1 }}
+      if: ${{ steps.p4est_cmake_serial_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
       run: rm -r $P4EST_SERIAL_DEBUG $P4EST_SERIAL_RELEASE || true 
     - name: make folders
