+++ .github/workflows/tests_t8code_cmake_linkage_parallel_debug.yml
@@ -106,7 +106,7 @@ jobs:
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       run: echo No cache found or cache will be ignored. SC_IGNORE_CACHE=$SC_IGNORE_CACHE
     - name: if ignore cache, delete folders
-      if: ${{ env.SC_IGNORE_CACHE == 1 }}
+      if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist
       run: rm -r $SC_DEBUG || true
     - name: make folders
@@ -152,10 +152,11 @@ jobs:
       run: echo "Install p4est"
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
     - name: if ignore cache, delete folders
-      if: ${{ env.P4EST_IGNORE_CACHE == 1 }}
+      if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
       run: rm -r $P4EST_DEBUG $P4EST_RELEASE || true 
     - name: make folders
+      if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       run: mkdir $P4EST_DEBUG $P4EST_RELEASE
     ## p4est debug
     - name: p4est cmake debug
