+++ .github/workflows/tests_t8code_automake_parallel.yml
@@ -100,7 +100,7 @@ jobs:
       if: ${{ steps.sc_automake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       run: echo No cache found or cache will be ignored. SC_IGNORE_CACHE=$SC_IGNORE_CACHE
     - name: if ignore cache, delete folders
-      if: ${{ env.SC_IGNORE_CACHE == 1 }}
+      if: ${{ steps.sc_automake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist
       run: rm -r $SC_DEBUG $SC_RELEASE || true
     - name: make folders
@@ -152,7 +152,7 @@ jobs:
       run: echo "Install p4est"
       if: ${{ steps.p4est_automake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
     - name: if ignore cache, delete folders
-      if: ${{ env.P4EST_IGNORE_CACHE == 1 }}
+      if: ${{ steps.p4est_automake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
       run: rm -r $P4EST_DEBUG $P4EST_RELEASE || true 
     - name: make folders
