### 2024-07-15 8f442a6
fixes in cmake CI
