### 2024-07-15 6c6027d
fixes in cmake CI
