### 2024-07-15 fd300c8
add cmake ci for parallel release with linkage
