+++ .github/workflows/tests_t8code_cmake_linkage_parallel_debug.yml
@@ -146,10 +146,10 @@ jobs:
     - name: if ignore cache, delete folders
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
-      run: rm -r $P4EST_DEBUG $P4EST_RELEASE || true 
+      run: rm -r $P4EST_DEBUG || true 
     - name: make folders
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
-      run: mkdir $P4EST_DEBUG $P4EST_RELEASE
+      run: mkdir $P4EST_DEBUG
     ## p4est debug
     - name: p4est cmake debug
       run: cd $P4EST_DEBUG && cmake ../ -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$PWD/install -DP4EST_USE_SYSTEM_SC=ON -DSC_DIR=$SC_DEBUG/install/cmake
