### 2024-07-15 DLR-AMR/t8code#1109 @maelk3
The combination OpenMPI together with libsc's CMake build system seems to be the culprit for the warnings. Libsc's CMake build system checks for the symbol `MPI_COMM_TYPE_SHARED` using the function `check_symbol_exists` in the header file `mpi.h.` MPICH defines this symbol as a macro but OpenMPI defines it as part of an anonymous enum which `check_symbol_exists` does not check for. Thus the compile definition `SC_ENABLE_MPICOMMSHARED` is missing causing the warnings.
