+++ .github/workflows/tests_t8code_cmake_linkage_parallel_release.yml
@@ -100,12 +100,19 @@ jobs:
     - name: if ignore cache, delete folders
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist
-      run: rm -r $SC_RELEASE || true
+      run: rm -r $SC_DEBUG $SC_RELEASE || true
     - name: make folders
-      run: mkdir $SC_RELEASE
+      run: mkdir $SC_DEBUG $SC_RELEASE
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
     - name: install sc
       run: echo "Install sc"
+      if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
+      ## sc debug
+    - name: sc cmake debug
+      run: cd $SC_DEBUG && cmake ../ -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$PWD/install -Dmpi=ON
+      if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
+    - name: sc build debug
+      run: cd $SC_DEBUG && make $MAKEFLAGS && make $MAKEFLAGS install
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
       ## sc release
     - name: sc cmake release
@@ -146,10 +153,10 @@ jobs:
     - name: if ignore cache, delete folders
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
-      run: rm -r $P4EST_RELEASE || true 
+      run: rm -r $SC_DEBUG $P4EST_RELEASE || true
     - name: make folders
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
-      run: mkdir $P4EST_RELEASE
+      run: mkdir $SC_DEBUG $P4EST_RELEASE
     ## p4est release
     - name: p4est cmake release
       run: cd $P4EST_RELEASE && cmake ../ -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/install -DP4EST_USE_SYSTEM_SC=ON -DSC_DIR=$SC_RELEASE/install/cmake
