+++ .github/workflows/tests_t8code_cmake_linkage_parallel_debug.yml
@@ -100,9 +100,9 @@ jobs:
     - name: if ignore cache, delete folders
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist
-      run: rm -r $SC_DEBUG || true
+      run: rm -r $SC_DEBUG $SC_RELEASE || true
     - name: make folders
-      run: mkdir $SC_DEBUG
+      run: mkdir $SC_DEBUG $SC_RELEASE
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
     - name: install sc
       run: echo "Install sc"
@@ -113,6 +113,13 @@ jobs:
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
     - name: sc build debug
       run: cd $SC_DEBUG && make $MAKEFLAGS && make $MAKEFLAGS install
+      if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
+      ## sc release
+    - name: sc cmake release
+      run: cd $SC_RELEASE && cmake ../ -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/install -Dmpi=ON
+      if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
+    - name: sc build release
+      run: cd $SC_RELEASE && make $MAKEFLAGS && make $MAKEFLAGS install
       if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' || env.SC_IGNORE_CACHE == 1 }}
 
 #
@@ -146,10 +153,10 @@ jobs:
     - name: if ignore cache, delete folders
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
       # The true at the end is to ignore errors that i.e. occur when the folders do not exist      
-      run: rm -r $P4EST_DEBUG || true 
+      run: rm -r $P4EST_DEBUG $P4EST_RELEASE || true
     - name: make folders
       if: ${{ steps.p4est_cmake_cache.outputs.cache-hit != 'true' || env.P4EST_IGNORE_CACHE == 1 }}
-      run: mkdir $P4EST_DEBUG
+      run: mkdir $P4EST_DEBUG $P4EST_RELEASE
     ## p4est debug
     - name: p4est cmake debug
       run: cd $P4EST_DEBUG && cmake ../ -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$PWD/install -DP4EST_USE_SYSTEM_SC=ON -DSC_DIR=$SC_DEBUG/install/cmake
