### 2024-07-15 fe5bff9
adapt test artifact path for cmake ci
