### 2024-07-15 5d99218
templated vtk writer to use the vtk API for cmeshes and forests
