### 2024-07-15 20c2450
Adding a CMake option.
