### 2024-07-16 69dd487
remove not yet implemented cmake less test option in ci
