+++ .github/workflows/tests_t8code_cmake_linkage_parallel_release.yml
@@ -179,7 +179,8 @@ jobs:
     # build config vars
     - name: less-test-option
       if: github.event_name == 'pull_request'
-      run: export LESS_TEST_OPTION="--enable-less-tests"
+      #placeholder for future implementation of the less test option
+      run: export LESS_TEST_OPTION=""
            && echo LESS_TEST_OPTION="$LESS_TEST_OPTION" >> $GITHUB_ENV
     - name: build config variables
       #TODO: Werror
