+++ .github/workflows/tests_t8code_cmake_parallel.yml
@@ -178,7 +178,8 @@ jobs:
     # build config vars
     - name: less-test-option
       if: github.event_name == 'pull_request'
-      run: export LESS_TEST_OPTION="--enable-less-tests"
+      #placeholder for future implementation of the less test option
+      run: export LESS_TEST_OPTION=""
            && echo LESS_TEST_OPTION="$LESS_TEST_OPTION" >> $GITHUB_ENV
     - name: build config variables
       #TODO: Werror
