+++ api/t8_fortran_interface/t8_fortran_interface.h
@@ -5,7 +5,6 @@
 
   Copyright (C) 2024 the developers
 
-
   t8code is free software; you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation; either version 2 of the License, or
