### 2024-07-16 d479feb
Fixed failing version test with CMake.
