### 2024-07-16 e592913
Add CMake option to enable less tests.
