### 2024-07-16 78288ac
Added netcdf cmake option
