### 2024-07-16 365f2f5
Update CMakeLists.txt
Co-authored-by: Sandro Elsweijer <49643115+sandro-elsweijer@users.noreply.github.com>
