### 2024-07-16 bb910b0
Add missing function to header
