### 2024-07-16 4dac49f
Add debug messages
