### 2024-07-17 2a8975a
Renamed inconsistent CMake build option
