+++ doc/CMakeLists.txt
@@ -1,4 +1,4 @@
-if(BUILD_DOCUMENTATION)
+if(T8CODE_BUILD_DOCUMENTATION)
     find_package(Doxygen REQUIRED)
     if (DOXYGEN_FOUND)
         set(DOXYGEN_OUTPUT_DIR ${PROJECT_BINARY_DIR}/doc)
@@ -29,9 +29,9 @@ if(BUILD_DOCUMENTATION)
     else (DOXYGEN_FOUND)
         message(FATAL_ERROR "Doxygen need to be installed to generate the doxygen documentation")
     endif (DOXYGEN_FOUND)
-endif(BUILD_DOCUMENTATION)
+endif(T8CODE_BUILD_DOCUMENTATION)
 
-if (BUILD_SPHINX_DOCUMENTATION)
+if (T8CODE_BUILD_SPHINX_DOCUMENTATION)
     set(CMAKE_MODULE_PATH "${PROJECT_SOURCE_DIR}/cmake" ${CMAKE_MODULE_PATH})
     find_package(Sphinx REQUIRED)
     
@@ -72,4 +72,4 @@ if (BUILD_SPHINX_DOCUMENTATION)
     else (SPHINX_FOUND)
         message(FATAL_ERROR "Sphinx need to be installed to generate the doxygen documentation")
     endif (SPHINX_FOUND)
-endif(BUILD_SPHINX_DOCUMENTATION)
+endif(T8CODE_BUILD_SPHINX_DOCUMENTATION)
