+++ CMakeLists.txt
@@ -22,8 +22,8 @@ option( T8CODE_ENABLE_OCC "Enable t8code's features which rely on OpenCASCADE" O
 option( T8CODE_USE_SYSTEM_SC "Use system-installed sc library" OFF )
 option( T8CODE_USE_SYSTEM_P4EST "Use system-installed p4est library" OFF )
 
-option( BUILD_DOCUMENTATION "Build t8code's documentation" ON)
-option( BUILD_SPHINX_DOCUMENTATION "Build t8code's documentation using sphinx" OFF)
+option( T8CODE_BUILD_DOCUMENTATION "Build t8code's documentation" ON)
+option( t8CODE_BUILD_SPHINX_DOCUMENTATION "Build t8code's documentation using sphinx" OFF)
 
 
 if( NOT CMAKE_BUILD_TYPE )
@@ -100,10 +100,10 @@ if ( T8CODE_BUILD_BENCHMARKS )
     add_subdirectory( ${CMAKE_CURRENT_LIST_DIR}/benchmarks )
 endif()
 
-if ( BUILD_DOCUMENTATION ) 
+if ( T8CODE_BUILD_DOCUMENTATION )
     add_subdirectory( ${CMAKE_CURRENT_LIST_DIR}/doc )
 endif()
 
-if ( BUILD_SPHINX_DOCUMENTATION )
-    set (BUILD_DOCUMENTATION 1)
+if ( T8CODE_BUILD_SPHINX_DOCUMENTATION )
+    set( T8CODE_BUILD_DOCUMENTATION 1 )
 endif()
