### 2024-07-17 031b353
adapt ci cmake linkage paths for occt and vtk
