### 2024-07-17 41f4599
enable less tests in cmake ci
