+++ .github/workflows/tests_t8code_linkage_parallel_debug.yml
@@ -210,7 +210,7 @@ jobs:
     - name: check OpenCASCADE
       run: echo "Checking OpenCASCADE with MPI in debugging mode"
     - name: configure MPI OpenCASCADE debug
-      run: mkdir build_cad && cd build_cad && ../configure $CONFIG_DEBUG --with-occ CFLAGS="$CFLAGS_var -I/usr/include/opencascade" CXXFLAGS="$CXXFLAGS_var -I/usr/include/opencascade"
+      run: mkdir build_cad && cd build_cad && ../configure $CONFIG_DEBUG --with-occ CFLAGS="$CFLAGS_var -I/usr/local/include/opencascade" CXXFLAGS="$CXXFLAGS_var -I/usr/local/include/opencascade"
     - name: OnFailUploadLog
       if: failure()
       uses: actions/upload-artifact@v4
@@ -218,11 +218,11 @@ jobs:
         name: config_cad_debug_MPI.log
         path: build_cad/config.log
     - name: make
-      run: cd build_cad && make $MAKEFLAGS CFLAGS="$CFLAGS_var -Werror -I/usr/include/opencascade" CXXFLAGS="$CXXFLAGS_var -Werror -I/usr/include/opencascade"
+      run: cd build_cad && make $MAKEFLAGS CFLAGS="$CFLAGS_var -Werror -I/usr/local/include/opencascade" CXXFLAGS="$CXXFLAGS_var -Werror -I/usr/local/include/opencascade"
     - name: make install
       run: cd build_cad && make install $MAKEFLAGS
     - name: make check
-      run: cd build_cad && make check $MAKEFLAGS CFLAGS="$CFLAGS_var -Werror -I/usr/include/opencascade" CXXFLAGS="$CXXFLAGS_var -Werror -I/usr/include/opencascade"
+      run: cd build_cad && make check $MAKEFLAGS CFLAGS="$CFLAGS_var -Werror -I/usr/local/include/opencascade" CXXFLAGS="$CXXFLAGS_var -Werror -I/usr/local/include/opencascade"
     - name: OnFailUploadLog
       if: failure()
       uses: actions/upload-artifact@v4
