### 2024-07-17 7b2db27
Switched off by default build documentation option in CMake
