### 2024-07-17 2280290
Remove function
