### 2024-07-17 fcccb4e
Renamed sphinx documentation CMake option
Co-authored-by: Johannes Markert <10619309+jmark@users.noreply.github.com>
