### 2024-07-17 0b34035
Add deleted docstring
