### 2024-07-17 9e76567
Introduce a c-interface for the vtk-writer
