### 2024-07-17 4ebed66
Fix sphinx cmake option variable name
