+++ doc/CMakeLists.txt
@@ -31,7 +31,7 @@ if(T8CODE_BUILD_DOCUMENTATION)
     endif (DOXYGEN_FOUND)
 endif(T8CODE_BUILD_DOCUMENTATION)
 
-if (T8CODE_BUILD_SPHINX_DOCUMENTATION)
+if (T8CODE_BUILD_DOCUMENTATION_SPHINX)
     set(CMAKE_MODULE_PATH "${PROJECT_SOURCE_DIR}/cmake" ${CMAKE_MODULE_PATH})
     find_package(Sphinx REQUIRED)
     
@@ -72,4 +72,4 @@ if (T8CODE_BUILD_SPHINX_DOCUMENTATION)
     else (SPHINX_FOUND)
         message(FATAL_ERROR "Sphinx need to be installed to generate the doxygen documentation")
     endif (SPHINX_FOUND)
-endif(T8CODE_BUILD_SPHINX_DOCUMENTATION)
+endif(T8CODE_BUILD_DOCUMENTATION_SPHINX)
