+++ CMakeLists.txt
@@ -104,6 +104,6 @@ if ( T8CODE_BUILD_DOCUMENTATION )
     add_subdirectory( ${CMAKE_CURRENT_LIST_DIR}/doc )
 endif()
 
-if ( T8CODE_BUILD_SPHINX_DOCUMENTATION )
+if ( T8CODE_BUILD_DOCUMENTATION_SPHINX )
     set( T8CODE_BUILD_DOCUMENTATION 1 )
 endif()
