### 2024-07-17 75dd2ca
Added note about meomry freeing for leaf_face_neighbors routine.
