### 2024-07-17 6a216f8
restructure first test
