+++ .github/workflows/tests_t8code_cmake_preparation.yml
@@ -1,0 +1,186 @@
+name: CMake CI preparation
+
+
+#  This file is part of t8code.
+#  t8code is a C library to manage a collection (a forest) of multiple
+#  connected adaptive space-trees of general element types in parallel.
+#
+#  Copyright (C) 2024 the developers
+#
+#  t8code is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation; either version 2 of the License, or
+#  (at your option) any later version.
+#
+#  t8code is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with t8code; if not, write to the Free Software Foundation, Inc.,
+#  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+
+#
+# This github CI script installs t8code and runs its tests for various configurations.
+# We compile sc and p4est as thirdparty libraries and use caching to only trigger a
+# new installation of them when their versions have changed in t8code.
+#
+# Note: To manually enforce sc and p4est installation, either increase the counter
+#       in the "key:" entries of the sc and p4est steps or set the variables
+#       SC_IGNORE_CACHE and P4EST_IGNORE_CACHE to 1 in the respective steps.
+
+on:
+  workflow_call:
+    inputs:
+      MAKEFLAGS:
+        required: true
+        type: string
+    outputs:
+      SC_DEBUG:
+        description: "Path to the debug build of sc"
+        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.SC_DEBUG }}
+      SC_RELEASE:
+        description: "Path to the release build of sc"
+        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.SC_RELEASE }}
+      P4EST_DEBUG:
+        description: "Path to the debug build of p4est"
+        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.P4EST_DEBUG }}
+      P4EST_RELEASE:
+        description: "Path to the release build of p4est"
+        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.P4EST_RELEASE }}
+
+jobs:
+  reusable_cmake_preparation_workflow:
+    runs-on: ubuntu-latest
+    container: dlramr/t8code-ubuntu:t8-dependencies
+    timeout-minutes: 20
+    outputs:
+      SC_DEBUG: ${{ steps.output_sc_folders.outputs.SC_DEBUG }}
+      SC_RELEASE: ${{ steps.output_sc_folders.outputs.SC_RELEASE }}
+      P4EST_DEBUG: ${{ steps.output_p4est_folders.outputs.P4EST_DEBUG }}
+      P4EST_RELEASE: ${{ steps.output_p4est_folders.outputs.P4EST_RELEASE }}
+    steps:
+#
+# Setup and bootstrap
+#
+    - uses: actions/checkout@v4
+      with: 
+        fetch-depth: 0
+    - name: Update packages
+      run: apt-get update && apt-get upgrade -y
[…]
