+++ .github/workflows/tests_t8code_cmake_parallel.yml
@@ -30,147 +30,30 @@ name: t8code cmake tests parallel
 #       in the "key:" entries of the sc and p4est steps or set the variables
 #       SC_IGNORE_CACHE and P4EST_IGNORE_CACHE to 1 in the respective steps.
 
-env:
-  MAKEFLAGS: "-j2"
-
 on:
-  push:
-    branches:
-      - main
-      - develop
-      - feature-*CI* # for testing this script, all feature branches with "CI" in their name
-  pull_request:
-    branches:
-      - main
-      - develop
-  workflow_dispatch: # Be able to trigger this manually on github.com
-  # Run every night at 1:10
-  schedule:
-      - cron:  '10 1 * * *'
+  workflow_call:
+    inputs:
+      MAKEFLAGS:
+        required: true
+        type: string
 
 jobs:
-  build:
-
-    if: (github.event_name == 'schedule' && github.repository == 'DLR-AMR/t8code') || (github.event_name != 'schedule')
-    runs-on: ubuntu-20.04
+  preparation:
+    uses: ./.github/workflows/tests_t8code_cmake_preparation.yml
+    with:
+      MAKEFLAGS: ${{ inputs.MAKEFLAGS }}
+  
+  cmake_tests_parallel:
+    needs: preparation
+    runs-on: ubuntu-latest
     container: dlramr/t8code-ubuntu:t8-dependencies
-    timeout-minutes: 90
+    timeout-minutes: 60
     steps:
-#
-# Setup and bootstrap
-#
-    - uses: actions/checkout@v4
-      with: 
-        fetch-depth: 0
-    - name: Update packages
-      run: apt-get update && apt-get upgrade -y
-      # This seems to be necessary because of the docker container
-    - name: disable ownership checks
-      run: git config --global --add safe.directory '*'
-    - name: init submodules
-      run: git submodule init
-    - name: update submodules
-      run: git submodule update
-#
-#       SC installation
-#
-    - name: store sc folders in var
-      run: echo SC_DEBUG=$PWD/sc/build_debug >> $GITHUB_ENV
-       && echo SC_RELEASE=$PWD/sc/build_release >> $GITHUB_ENV
-    - name: Get sc commit hash
-      run: hash=`git rev-parse HEAD:sc` && echo sc_commit=$hash >> $GITHUB_ENV
-    - name: Check cache for previous sc installation
-      id: sc_cmake_cache
-      uses: actions/cache@v4
-      with:
-        path: |
-          ${{ env.SC_DEBUG }}
-          ${{ env.SC_RELEASE }}
-        # You can increase the counter at the end to force a new key and hence recomputing the cache
-        key: sc-cmake-${{ env.sc_commit }}-005
-    - name: Set ignore cache variable
-      # If this variable is set 1
-      # (i.e. SC_IGNORE_CACHE=1) then the cache will be ignored and
-      # sc is always build.
-      # We use this mostly for debugging this CI script.
-      run: echo SC_IGNORE_CACHE=0 >> $GITHUB_ENV
-    - name: Cache info
-      if: ${{ steps.sc_cmake_cache.outputs.cache-hit != 'true' ||  env.SC_IGNORE_CACHE == 1 }}
[…]
