+++ .github/workflows/test_tarball.yml
@@ -43,7 +43,7 @@ on:
 jobs:
   build:
 
-    if: (github.event_name == 'schedule' && github.repository == 'DLR-AMR/t8code') || (github.event_name != 'schedule')
+    if: 0
     runs-on: ubuntu-20.04
     container: dlramr/t8code-ubuntu:t8-dependencies
     timeout-minutes: 90
