### 2024-07-17 e109625
Change attribute offset computation of ghost attributes when commiting from stash
