+++ src/t8_cmesh/t8_cmesh_trees.h
@@ -220,7 +220,7 @@ t8_cmesh_trees_start_part (t8_cmesh_trees_t trees, int proc, t8_locidx_t lfirst_
  * \param [in,out]        trees The trees structure to be updated.
  * \param [in]            proc  The number of the part to be finished.
  */
-size_t
+void
 t8_cmesh_trees_finish_part (t8_cmesh_trees_t trees, int proc);
 
 /** Copy the tree_to_proc and ghost_to_proc arrays of one tree structure to
@@ -435,7 +435,7 @@ t8_cmesh_trees_add_attribute (const t8_cmesh_trees_t trees, int proc, const t8_s
 
 void
 t8_cmesh_trees_add_ghost_attribute (const t8_cmesh_trees_t trees, int proc, const t8_stash_attribute_struct_t *attr,
-                                    t8_locidx_t local_ghost_id, size_t index, size_t *attribute_data_offset);
+                                    t8_locidx_t local_ghost_id, size_t index);
 
 /** Return the number of parts of a trees structure.
  * \param [in]        trees The trees structure.
