### 2024-07-17 5171e7b
Switched sphinx documentation to dependent option
