### 2024-07-17 6d34f00
specify ressources only in topmost script
