+++ .github/workflows/tests_t8code.yml
@@ -48,6 +48,8 @@ on:
 jobs:
   cmake_tests_parallel:
     if: (github.event_name == 'schedule' && github.repository == 'DLR-AMR/t8code') || (github.event_name != 'schedule')
+    runs-on: ubuntu-latest
+    container: dlramr/t8code-ubuntu:t8-dependencies
     uses: ./.github/workflows/tests_t8code_cmake_parallel.yml
     with:
       MAKEFLAGS: -j2
