+++ .github/workflows/tests_t8code.yml
@@ -1,4 +1,4 @@
-name: t8code cmake tests serial
+name: t8code tests
 
 
 #  This file is part of t8code.
