### 2024-07-17 21ec04b
Finalize Interface
