### 2024-07-17 a61378b
CMake pedantic build mode option.
