### 2024-07-17 fd7f0e5
Swiched to CXX flags naming.
