+++ src/CMakeLists.txt
@@ -59,6 +59,7 @@ endif()
 
 if( T8CODE_BUILD_PEDANTIC )
   # TODO: Add `-Wextra -Wpedantic`. As of now, t8code would not compile with this.
+  # See also: https://github.com/DLR-AMR/t8code/issues/1155.
   target_compile_options( T8 PUBLIC -Wall -pedantic -Werror )
   set (T8_CXXFLAGS "${T8_CXXFLAGS} -W -Wall -Wpedantic -Werror")
   set (T8_CFLAGS "${T8_CFLAGS} -W -Wall -Wpedantic -Werror")
