### 2024-07-17 663ce2f
Remove old functions
