+++ src/t8_cmesh/t8_cmesh_vtk_writer.c
@@ -19,292 +19,3 @@
   along with t8code; if not, write to the Free Software Foundation, Inc.,
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
-
-#include <t8_cmesh_vtk_writer.h>
-#include <t8_vtk.h>
-#include "t8_cmesh_trees.h"
-#include "t8_cmesh_types.h"
-
-/* Return the local number of vertices in a cmesh.
- * \param [in] cmesh       The cmesh to be considered.
- * \param [in] count_ghosts If true, we also count the vertices of the ghost trees.
- * \return                 The number of vertices associated to \a cmesh.
- * \a cmesh must be committed before calling this function.
- */
-static t8_gloidx_t
-t8_cmesh_get_num_vertices (const t8_cmesh_t cmesh, const int count_ghosts)
-{
-  int iclass;
-  t8_eclass_t ghost_class;
-  t8_gloidx_t num_vertices = 0;
-  t8_locidx_t ighost;
-  T8_ASSERT (cmesh != NULL);
-  T8_ASSERT (cmesh->committed);
-
-  for (iclass = T8_ECLASS_ZERO; iclass < T8_ECLASS_COUNT; iclass++) {
-    num_vertices += t8_eclass_num_vertices[iclass] * cmesh->num_local_trees_per_eclass[iclass];
-  }
-  if (count_ghosts) {
-    /* Also count the vertices of the ghost trees */
-    for (ighost = 0; ighost < t8_cmesh_get_num_ghosts (cmesh); ighost++) {
-      ghost_class = t8_cmesh_get_ghost_class (cmesh, ighost);
-      num_vertices += t8_eclass_num_vertices[ghost_class];
-    }
-  }
-
-  return num_vertices;
-}
-
-static int
-t8_cmesh_vtk_write_file_ext (const t8_cmesh_t cmesh, const char *fileprefix, const int write_ghosts)
-{
-  T8_ASSERT (cmesh != NULL);
-  T8_ASSERT (t8_cmesh_is_committed (cmesh));
-  T8_ASSERT (fileprefix != NULL);
-
-  if (cmesh->mpirank == 0) {
-    /* Write the pvtu header file. */
-    int num_ranks_that_write = cmesh->set_partition ? cmesh->mpisize : 1;
-    if (t8_write_pvtu (fileprefix, num_ranks_that_write, 1, 1, 0, 0, 0, NULL)) {
-      SC_ABORTF ("Error when writing file %s.pvtu\n", fileprefix);
-    }
-  }
-  /* If the cmesh is replicated only rank 0 prints it,
-   * otherwise each process prints its part of the cmesh.*/
-  if (cmesh->mpirank == 0 || cmesh->set_partition) {
-    char vtufilename[BUFSIZ];
-    FILE *vtufile;
-    t8_locidx_t num_vertices, ivertex;
-    t8_locidx_t num_trees;
-    t8_ctree_t tree;
-    double x, y, z;
-    double *vertices, *vertex;
-    int k, sk;
-    long long offset, count_vertices;
-    t8_locidx_t ighost, num_ghosts = 0, num_loc_trees;
-#ifdef T8_ENABLE_DEBUG
-    t8_cghost_t ghost;
-#endif
-    t8_eclass_t eclass;
-
-    num_vertices = t8_cmesh_get_num_vertices (cmesh, write_ghosts);
-    num_trees = t8_cmesh_get_num_local_trees (cmesh);
-    if (write_ghosts) {
-      num_trees += t8_cmesh_get_num_ghosts (cmesh);
-    }
-
[…]
