### 2024-07-18 4ebab16
use pedantic build mode in cmake ci
