+++ .github/workflows/tests_t8code_cmake_serial.yml
@@ -181,8 +181,7 @@ jobs:
       run: export LESS_TEST_OPTION="-DT8CODE_ENABLE_LESS_TESTS=ON"
            && echo LESS_TEST_OPTION="$LESS_TEST_OPTION" >> $GITHUB_ENV
     - name: build config variables
-      #TODO: Werror
-      run: export CONFIG_OPTIONS="${LESS_TEST_OPTION} -DT8CODE_USE_SYSTEM_SC=ON -DT8CODE_USE_SYSTEM_P4EST=ON -DT8CODE_ENABLE_MPI=OFF"
+      run: export CONFIG_OPTIONS="${LESS_TEST_OPTION} -DT8CODE_USE_SYSTEM_SC=ON -DT8CODE_USE_SYSTEM_P4EST=ON -DT8CODE_ENABLE_MPI=OFF -DT8CODE_BUILD_PEDANTIC=ON"
            && export CONFIG_SERIAL_DEBUG="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Debug -DSC_DIR=$SC_SERIAL_DEBUG/install/cmake -DP4EST_DIR=$P4EST_SERIAL_DEBUG/install/cmake"
            && export CONFIG_SERIAL_RELEASE="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Release -DSC_DIR=$SC_SERIAL_RELEASE/install/cmake -DP4EST_DIR=$P4EST_SERIAL_RELEASE/install/cmake"
            && echo CONFIG_OPTIONS="$CONFIG_OPTIONS" >> $GITHUB_ENV
