### 2024-07-18 DLR-AMR/t8code#1160 @jwallwork23
"UNKNOWN" heading in documentation
The title on page https://dlr-amr.github.io/t8code/doc/v2.0.0/index.html is "t8", followed by "UNKNOWN". Similarly for other versions and "latest". However, on the https://dlr-amr.github.io/t8code/doc/dev/index.html page there is a version number here (`2.0.0.496-6683`). Is there an issue with the pages for versions other than "dev"?

(Part of JOSS review https://github.com/openjournals/joss-reviews/issues/6887.)
