+++ benchmarks/time_new_refine.c
@@ -21,9 +21,11 @@
 */
 
 #include <sc_refcount.h>
-#include <t8_schemes/t8_default/t8_dtri.h>
-#include <t8_schemes/t8_default/t8_dtet.h>
-#include <t8_schemes/t8_default.h>
+#include <t8_cmesh/t8_cmesh_examples.h>
+#include <t8_element_c_interface.h>
+#include <t8_schemes/t8_default/t8_default_c_interface.h>
+#include <t8_schemes/t8_default/t8_default_tri/t8_dtri.h>
+#include <t8_schemes/t8_default/t8_default_tet/t8_dtet.h>
 #include <t8_forest/t8_forest_adapt.h>
 #include <t8_forest/t8_forest_general.h>
 #include <sc_flops.h>
