### 2024-07-18 83cfb34
try using artifacts in cmake ci
