### 2024-07-18 DLR-AMR/t8code#1161 @jwallwork23
More detail on contributing
The "Contributing" section at the start of `CONTRIBUTING.md` is very brief. It would be beneficial to provide a little more detail on:
 a) How to contact principal developers (if you want to make a large contribution).
 b) Provide a link to the coding guidelines [wiki page](https://github.com/DLR-AMR/t8code/wiki/Coding-Guideline).
 c) Whether there are requirements for all new contributions to be tested and documented. I notice there is a wiki page on this but it isn't linked from `CONTRIBUTING.md` nor the coding guidelines [wiki page](https://github.com/DLR-AMR/t8code/wiki/Coding-Guideline).

(Part of JOSS review https://github.com/openjournals/joss-reviews/issues/6887.)
