### 2024-07-18 fc7e07e
Enabled previously disabled code
