### 2024-07-18 fc97681
disable ownership checks in cmake ci
