+++ .github/workflows/tests_t8code_cmake_parallel.yml
@@ -58,6 +58,11 @@ jobs:
       uses: actions/download-artifact@v4
       with:
         name: SC_P4EST_MPI
+    - name: Update packages
+      run: apt-get update && apt-get upgrade -y
+      # This seems to be necessary because of the docker container
+    - name: disable ownership checks
+      run: git config --global --add safe.directory '*'
 #
 #  T8CODE
 #
