### 2024-07-18 d5a8f2a
Added benchmark into both build systems
