### 2024-07-18 532fca7
Rename T8_TREE/GHOST_FIRST_ATT to T8_TREE/GHOST_FIRST_ATT_INFO
