+++ src/t8_cmesh/t8_cmesh_trees.h
@@ -121,14 +121,14 @@ T8_EXTERN_C_BEGIN ();
 /* TODO: document */
 
 /* Given a tree return the beginning of its attributes block */
-#define T8_TREE_FIRST_ATT(t) ((char *) (t) + (t)->att_offset)
+#define T8_TREE_FIRST_ATT_INFO (t) ((char *) (t) + (t)->att_offset)
 
 /* Given a tree and an index i return the i-th attribute index of that tree */
 #define T8_TREE_ATTR_INFO(t, i) \
   ((t8_attribute_info_struct_t *) ((char *) (t) + (t)->att_offset + (i) * sizeof (t8_attribute_info_struct_t)))
 
 /* Given a tree and an attribute info return the attribute */
-#define T8_TREE_ATTR(t, ai) (T8_TREE_FIRST_ATT (t) + (ai)->attribute_offset)
+#define T8_TREE_ATTR(t, ai) (T8_TREE_FIRST_ATT_INFO (t) + (ai)->attribute_offset)
 
 /* Given a tree return its face_neighbor array */
 #define T8_TREE_FACE(t) ((char *) (t) + (t)->neigh_offset)
@@ -137,7 +137,7 @@ T8_EXTERN_C_BEGIN ();
 #define T8_TREE_TTF(t) (T8_TREE_FACE (t) + t8_eclass_num_faces[(t)->eclass] * sizeof (t8_locidx_t))
 
 /* Given a ghost return the beginning of its attribute block */
-#define T8_GHOST_FIRST_ATT(g) T8_TREE_FIRST_ATT (g)
+#define T8_GHOST_FIRST_ATT_INFO (g) T8_TREE_FIRST_ATT_INFO (g)
 
 /* Given a ghost and an index i return the i-th attribute index of that ghost */
 #define T8_GHOST_ATTR_INFO(g, i) T8_TREE_ATTR_INFO (g, i)
