### 2024-07-18 1bc5758
further restructuring of cmake ci
