+++ .github/workflows/tests_t8code_cmake_preparation_parallel.yml
@@ -1,4 +1,4 @@
-name: CMake CI preparation
+name: Parallel CMake CI preparation
 
 
 #  This file is part of t8code.
@@ -21,48 +21,54 @@ name: CMake CI preparation
 #  along with t8code; if not, write to the Free Software Foundation, Inc.,
 #  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-#
-# This github CI script installs t8code and runs its tests for various configurations.
-# We compile sc and p4est as thirdparty libraries and use caching to only trigger a
-# new installation of them when their versions have changed in t8code.
-#
-# Note: To manually enforce sc and p4est installation, either increase the counter
-#       in the "key:" entries of the sc and p4est steps or set the variables
-#       SC_IGNORE_CACHE and P4EST_IGNORE_CACHE to 1 in the respective steps.
-
 on:
   workflow_call:
     inputs:
       MAKEFLAGS:
         required: true
         type: string
+      IGNORE_CACHE:
+        required: false
+        type: boolean
+        default: false
+      CACHE_COUNTER:
+        required: true
+        type: string
     outputs:
       SC_DEBUG:
         description: "Path to the debug build of sc"
-        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.SC_DEBUG }}
+        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.SC_DEBUG }}
       SC_RELEASE:
         description: "Path to the release build of sc"
-        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.SC_RELEASE }}
+        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.SC_RELEASE }}
       P4EST_DEBUG:
         description: "Path to the debug build of p4est"
-        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.P4EST_DEBUG }}
+        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.P4EST_DEBUG }}
       P4EST_RELEASE:
         description: "Path to the release build of p4est"
-        value: ${{ jobs.reusable_cmake_preparation_workflow.outputs.P4EST_RELEASE }}
+        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.P4EST_RELEASE }}
+      SC_CACHE_HIT:
+        description: "Whether the sc cache was hit"
+        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.SC_CACHE_HIT }}
+      P4EST_CACHE_HIT:
+        description: "Whether the p4est cache was hit"
+        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.P4EST_CACHE_HIT }}
 
 jobs:
-  reusable_cmake_preparation_workflow:
+  reusable_cmake_preparation_workflow_parallel:
     runs-on: ubuntu-latest
     container: dlramr/t8code-ubuntu:t8-dependencies
-    timeout-minutes: 20
+    timeout-minutes: 10
     outputs:
       SC_DEBUG: ${{ steps.output_sc_folders.outputs.SC_DEBUG }}
       SC_RELEASE: ${{ steps.output_sc_folders.outputs.SC_RELEASE }}
       P4EST_DEBUG: ${{ steps.output_p4est_folders.outputs.P4EST_DEBUG }}
       P4EST_RELEASE: ${{ steps.output_p4est_folders.outputs.P4EST_RELEASE }}
+      SC_CACHE_HIT: ${{ steps.sc_cmake_cache.outputs.cache-hit }}
[…]
