+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,39 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug/unexpected behaviour is.
+
+**To Reproduce**
+Which command did you run/which function did you call?
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Log files**
+- Please upload the `config.log` file
+- For failing tests, please upload the `test-suite.log` file
+
+**Environment (please complete the following information):**
+ - OS: [e.g. Ubuntu 22.04, Windows 11]
+ - Compiler: [e.g., gcc, icc, mpicc]
+ - Compiler version: [e.g. the output of 'gcc --version']
+ - MPI version (if MPI is used): [The output of 'mpirun --version' or 'mpiexec --version']
+ 
+**Estimated priority**
+Which of these is most applicable (remove the others):
+
+"Priority: high"  Should be solved as soon as possible 
+
+"Priority: medium"  Should be solved within half a year 
+
+"Priority: low" Should be solved eventually
+
+**Additional context**
+Add any other context about the problem here.
