### 2024-07-18 f84f90d
Update issue templates
