+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,30 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Estimated priority**
+Which of these is most applicable (remove the others):
+
+"Priority: high"  Should be solved as soon as possible 
+
+"Priority: medium"  Should be solved within half a year 
+
+"Priority: low" Should be solved eventually
+
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
