+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,10 @@
+---
+name: Question
+about: I have a question on how to use t8code
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+Please post any questions about t8code or the usage of t8code.
