### 2024-07-18 6dcd90a
Update issue templates
