+++ .github/ISSUE_TEMPLATE/question.md
@@ -7,4 +7,6 @@ assignees: ''
 
 ---
 
+# Question
+
 Please post any questions about t8code or the usage of t8code.
