+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,6 +7,8 @@ assignees: ''
 
 ---
 
+# Bug report
+
 **Describe the bug**
 A clear and concise description of what the bug/unexpected behaviour is.
 
