### 2024-07-18 7666f6f
further restructuring of cmake ci
