+++ .github/workflows/tests_t8code_cmake_preparation_serial.yml
@@ -1,193 +1,0 @@
-name: Serial CMake CI preparation
-
-
-#  This file is part of t8code.
-#  t8code is a C library to manage a collection (a forest) of multiple
-#  connected adaptive space-trees of general element types in parallel.
-#
-#  Copyright (C) 2024 the developers
-#
-#  t8code is free software; you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation; either version 2 of the License, or
-#  (at your option) any later version.
-#
-#  t8code is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with t8code; if not, write to the Free Software Foundation, Inc.,
-#  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-
-on:
-  workflow_call:
-    inputs:
-      MAKEFLAGS:
-        required: true
-        type: string
-      IGNORE_CACHE:
-        required: false
-        type: boolean
-        default: false
-      CACHE_COUNTER:
-        required: true
-        type: string
-    outputs:
-      SC_SERIAL_DEBUG:
-        description: "Path to the debug build of sc"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_serial.outputs.SC_SERIAL_DEBUG }}
-      SC_SERIAL_RELEASE:
-        description: "Path to the release build of sc"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_serial.outputs.SC_SERIAL_RELEASE }}
-      P4EST_SERIAL_DEBUG:
-        description: "Path to the debug build of p4est"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_serial.outputs.P4EST_SERIAL_DEBUG }}
-      P4EST_SERIAL_RELEASE:
-        description: "Path to the release build of p4est"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_serial.outputs.P4EST_SERIAL_RELEASE }}
-      SC_SERIAL_CACHE_HIT:
-        description: "Whether the sc cache was hit"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_serial.outputs.SC_SERIAL_CACHE_HIT }}
-      P4EST_SERIAL_CACHE_HIT:
-        description: "Whether the p4est cache was hit"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_serial.outputs.P4EST_SERIAL_CACHE_HIT }}
-
-jobs:
-  reusable_cmake_preparation_workflow_serial:
-    runs-on: ubuntu-latest
-    container: dlramr/t8code-ubuntu:t8-dependencies
-    timeout-minutes: 10
-    outputs:
-      SC_SERIAL_DEBUG: ${{ steps.output_sc_folders.outputs.SC_SERIAL_DEBUG }}
-      SC_SERIAL_RELEASE: ${{ steps.output_sc_folders.outputs.SC_SERIAL_RELEASE }}
-      P4EST_SERIAL_DEBUG: ${{ steps.output_p4est_folders.outputs.P4EST_SERIAL_DEBUG }}
[…]
