+++ .github/workflows/tests_t8code_cmake_preparation.yml
@@ -1,4 +1,4 @@
-name: Parallel CMake CI preparation
+name: CMake CI preparation
 
 
 #  This file is part of t8code.
@@ -34,28 +34,31 @@ on:
       CACHE_COUNTER:
         required: true
         type: string
+      MPI:
+        required: true
+        type: string
     outputs:
       SC_DEBUG:
         description: "Path to the debug build of sc"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.SC_DEBUG }}
+        value: ${{ jobs.cmake_preparation.outputs.SC_DEBUG }}
       SC_RELEASE:
         description: "Path to the release build of sc"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.SC_RELEASE }}
+        value: ${{ jobs.cmake_preparation.outputs.SC_RELEASE }}
       P4EST_DEBUG:
         description: "Path to the debug build of p4est"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.P4EST_DEBUG }}
+        value: ${{ jobs.cmake_preparation.outputs.P4EST_DEBUG }}
       P4EST_RELEASE:
         description: "Path to the release build of p4est"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.P4EST_RELEASE }}
-      SC_CACHE_HIT:
-        description: "Whether the sc cache was hit"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.SC_CACHE_HIT }}
-      P4EST_CACHE_HIT:
-        description: "Whether the p4est cache was hit"
-        value: ${{ jobs.reusable_cmake_preparation_workflow_parallel.outputs.P4EST_CACHE_HIT }}
+        value: ${{ jobs.cmake_preparation.outputs.P4EST_RELEASE }}
+      USED_CACHE:
+        description: "Whether the cache was used"
+        value: ${{ jobs.cmake_preparation.outputs.USED_CACHE }}
+
+env:
+  USED_CACHE: $IGNORE_CACHE
 
 jobs:
-  reusable_cmake_preparation_workflow_parallel:
+  cmake_preparation:
     runs-on: ubuntu-latest
     container: dlramr/t8code-ubuntu:t8-dependencies
     timeout-minutes: 10
@@ -64,8 +67,7 @@ jobs:
       SC_RELEASE: ${{ steps.output_sc_folders.outputs.SC_RELEASE }}
       P4EST_DEBUG: ${{ steps.output_p4est_folders.outputs.P4EST_DEBUG }}
       P4EST_RELEASE: ${{ steps.output_p4est_folders.outputs.P4EST_RELEASE }}
-      SC_CACHE_HIT: ${{ steps.sc_cmake_cache.outputs.cache-hit }}
-      P4EST_CACHE_HIT: ${{ steps.p4est_cmake_cache.outputs.cache-hit }}
+      USED_CACHE: ${{ steps.used_cache.outputs.USED_CACHE }}
     steps:
 #
 # Setup
@@ -103,33 +105,36 @@ jobs:
           ${{ env.SC_RELEASE }}
         # You can increase the counter at the end to force a new key and hence recomputing the cache
         key: sc-cmake-${{ env.sc_commit }}-${{ env.CACHE_COUNTER }}
[…]
