+++ .github/workflows/tests_t8code_cmake_tests.yml
@@ -1,4 +1,4 @@
-name: t8code serial CMake tests
+name: CMake CI t8code tests
 
 
 #  This file is part of t8code.
@@ -27,6 +27,9 @@ on:
       MAKEFLAGS:
         required: true
         type: string
+      MPI:
+        required: true
+        type: string
       SC_DEBUG:
         required: true
         type: string
@@ -41,7 +44,7 @@ on:
         type: string
 
 jobs:  
-  cmake_tests_parallel:
+  t8code_cmake_tests:
     timeout-minutes: 30
     runs-on: ubuntu-latest
     container: dlramr/t8code-ubuntu:t8-dependencies
@@ -52,7 +55,7 @@ jobs:
     - name: Download artifacts
       uses: actions/download-artifact@v4
       with:
-        name: SC_P4EST_PARALLEL
+        name: SC_P4EST_MPI_${{ env.MPI }}
     - name: Update packages
       run: apt-get update && apt-get upgrade -y
       # This seems to be necessary because of the docker container

[…]
