### 2024-07-18 f9b91a9
further restructuring of cmake ci
