+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -19,8 +19,12 @@ Which command did you run/which function did you call?
 A clear and concise description of what you expected to happen.
 
 **Log files**
-- Please upload the `config.log` file
-- For failing tests, please upload the `test-suite.log` file
+- Please upload the `config.log` file (when building with autotools)
+- For failing tests, please upload the `test-suite.log` file (when building with autotools)
+- When building with CMake please upload
+  - `CMakeCache.txt`
+  - `CMakeFiles/CMakeError.log`
+  - `CMakeFiles/CMakeOutput.log`
 
 **Environment (please complete the following information):**
  - OS: [e.g. Ubuntu 22.04, Windows 11]
