### 2024-07-18 1803e2b
further restructuring of cmake ci
