+++ .github/workflows/tests_sc_p4est_cmake_tests.yml
@@ -75,7 +75,7 @@ jobs:
 #
       ## sc debug
     - name: sc debug check
-      run: cd $SC_DEBUG && make test $MAKEFLAGS
+      run: cd $SC_DEBUG && ctest $MAKEFLAGS
     - name: OnFailUploadLog
       if: failure()
       uses: actions/upload-artifact@v4
@@ -84,7 +84,7 @@ jobs:
         path: $SC_DEBUG/Testing/Temporary/LastTest.log
       ## sc release
     - name: sc release check
-      run: cd $SC_RELEASE && make test $MAKEFLAGS
+      run: cd $SC_RELEASE && ctest $MAKEFLAGS
     - name: OnFailUploadLog
       if: failure()
       uses: actions/upload-artifact@v4
@@ -96,7 +96,7 @@ jobs:
 #
       ## p4est debug
     - name: p4est debug check
-      run: cd $P4EST_DEBUG && make test $MAKEFLAGS
+      run: cd $P4EST_DEBUG && ctest $MAKEFLAGS
     - name: OnFailUploadLog
       if: failure()
       uses: actions/upload-artifact@v4
@@ -105,7 +105,7 @@ jobs:
         path: $P4EST_DEBUG/Testing/Temporary/LastTest.log
       ## p4est release
     - name: p4est release check
-      run: cd $P4EST_RELEASE && make test $MAKEFLAGS
+      run: cd $P4EST_RELEASE && ctest $MAKEFLAGS
     - name: OnFailUploadLog
       if: failure()
       uses: actions/upload-artifact@v4
