+++ .github/workflows/tests_t8code_cmake_tests.yml
@@ -61,6 +61,17 @@ jobs:
       # This seems to be necessary because of the docker container
     - name: disable ownership checks
       run: git config --global --add safe.directory '*'
+    - name: Get input vars
+      run: export MAKEFLAGS="${{ inputs.MAKEFLAGS }}"
+        && export SC_DEBUG="${{ inputs.SC_DEBUG }}"
+        && export SC_RELEASE="${{ inputs.SC_RELEASE }}"
+        && export P4EST_DEBUG="${{ inputs.P4EST_DEBUG }}"
+        && export P4EST_RELEASE="${{ inputs.P4EST_RELEASE }}"
+        && echo MAKEFLAGS="$MAKEFLAGS" >> $GITHUB_ENV
+        && echo SC_DEBUG="$SC_DEBUG" >> $GITHUB_ENV
+        && echo SC_RELEASE="$SC_RELEASE" >> $GITHUB_ENV
+        && echo P4EST_DEBUG="$P4EST_DEBUG" >> $GITHUB_ENV
+        && echo P4EST_RELEASE="$P4EST_RELEASE" >> $GITHUB_ENV
 #
 #  T8CODE
 #
@@ -68,33 +79,33 @@ jobs:
     - name: less-test-option
       if: github.event_name == 'pull_request'
       run: export LESS_TEST_OPTION="-DT8CODE_ENABLE_LESS_TESTS=ON"
-           && echo LESS_TEST_OPTION="$LESS_TEST_OPTION" >> $GITHUB_ENV
+        && echo LESS_TEST_OPTION="$LESS_TEST_OPTION" >> $GITHUB_ENV
     - name: build config variables
       run: export CONFIG_OPTIONS="${LESS_TEST_OPTION} -DT8CODE_USE_SYSTEM_SC=ON -DT8CODE_USE_SYSTEM_P4EST=ON -DT8CODE_BUILD_PEDANTIC=ON"
-           && export CONFIG_DEBUG="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Debug -DSC_DIR=$SC_DEBUG/install/cmake -DP4EST_DIR=$P4EST_DEBUG/install/cmake"
-           && export CONFIG_RELEASE="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Release -DSC_DIR=$SC_RELEASE/install/cmake -DP4EST_DIR=$P4EST_RELEASE/install/cmake"
-           && echo CONFIG_OPTIONS="$CONFIG_OPTIONS" >> $GITHUB_ENV
-           && echo CONFIG_DEBUG="$CONFIG_DEBUG" >> $GITHUB_ENV
-           && echo CONFIG_RELEASE="$CONFIG_RELEASE" >> $GITHUB_ENV
+        && export CONFIG_DEBUG="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Debug -DSC_DIR=$SC_DEBUG/install/cmake -DP4EST_DIR=$P4EST_DEBUG/install/cmake"
+        && export CONFIG_RELEASE="$CONFIG_OPTIONS -DCMAKE_BUILD_TYPE=Release -DSC_DIR=/install/cmake -DP4EST_DIR=$P4EST_RELEASE/install/cmake"
+        && echo CONFIG_OPTIONS="$CONFIG_OPTIONS" >> $GITHUB_ENV
+        && echo CONFIG_DEBUG="$CONFIG_DEBUG" >> $GITHUB_ENV
+        && echo CONFIG_RELEASE="$CONFIG_RELEASE" >> $GITHUB_ENV
     - name: Check vars
       run: echo "[$CONFIG_DEBUG]"
-            && echo "[$CONFIG_RELEASE]"
-# configure and test with MPI (release mode)
[…]
