### 2024-07-18 5393456
further restructuring of cmake ci
