### 2024-07-18 ca9da4e
debugging cmake CI
