+++ .github/workflows/tests_t8code.yml
@@ -52,7 +52,7 @@ jobs:
     uses: ./.github/workflows/tests_t8code_cmake_preparation.yml
     with:
       MAKEFLAGS: -j4
-      IGNORE_CACHE: false
+      IGNORE_CACHE: true
       CACHE_COUNTER: 5
       MPI: OFF
   
@@ -89,7 +89,7 @@ jobs:
     uses: ./.github/workflows/tests_t8code_cmake_preparation.yml
     with:
       MAKEFLAGS: -j4
-      IGNORE_CACHE: false
+      IGNORE_CACHE: true
       CACHE_COUNTER: 5
       MPI: ON
   
