### 2024-07-18 b8630dc
further restructuring of cmake ci
