### 2024-07-18 b82b6bc
Fix compilation without vtk
