### 2024-07-18 541d65e
further restructuring of cmake ci
