### 2024-07-19 DLR-AMR/t8code#1160 @holke
@jwallwork23 thank you for reporting.

This is a duplicate of https://github.com/DLR-AMR/t8code/issues/1070.
Basically there was an issue in our CI that generates the documentation that resulted in the version not being visible.

The issue is resolved and upcoming versions will have the  correct version number.
