### 2024-07-19 DLR-AMR/t8code#1161 @holke
Thank you for this input.

It is very beneficial for us to have a good CONTRIBUTING file.
We will adress your changes.
