### 2024-07-19 830705f
rename mattermost jobs to be more precise
