+++ .github/workflows/mattermost issue.yml
@@ -23,7 +23,7 @@
 # See https://docs.github.com/en/webhooks/webhook-events-and-payloads?actionType=closed#issues for
 # a documentation of the github issues event.
 
-name: Mattermost_message
+name: Mattermost_message_issue
 
 # Controls when the workflow will run
 on:
@@ -37,7 +37,7 @@ env:
   message_build: 0
   
 jobs:
-  send_mm_message:
+  send_mm_message_issue:
     if: github.repository == 'DLR-AMR/t8code'
     runs-on: ubuntu-20.04
     steps:
