### 2024-07-19 32ee96d
use github.event.action to get type
