### 2024-07-19 e0d09e9
remove sudo in cmake ci
