+++ .github/workflows/tests_t8code_cmake_tests.yml
@@ -56,6 +56,8 @@ jobs:
       uses: actions/download-artifact@v4
       with:
         name: SC_P4EST_MPI_${{ inputs.MPI }}
+    - name: untar artifact
+      run: tar -xf artifact.tar && rm artifact.tar
     - name: Update packages
       run: apt-get update && apt-get upgrade -y
       # This seems to be necessary because of the docker container
