+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -12,7 +12,7 @@ assignees: ''
 **Is your feature request related to a problem? Please describe.**
 A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
 
-**Describe the solution you'd like**
+**Describe the solution or feature you'd like**
 A clear and concise description of what you want to happen.
 
 **Describe alternatives you've considered**
