### 2024-07-19 0e921a6
Fixes for offset calculation
