### 2024-07-19 4da7c63
Move template specializations into seperate cxx file to avoid linker error
Linker found multiple definitions
