### 2024-07-19 c4a5fbc
Update src/CMakeLists.txt
Co-authored-by: Johannes Markert <10619309+jmark@users.noreply.github.com>
