### 2024-07-19 DLR-AMR/t8code#1161 @holke
Please see #1172. I hope i addressed all your concernces.
I added further info, linked the Wiki article on Coding guidelines and also updated the Wiki article.
