### 2024-07-20 #6887 @elbeejay
Hi everyone, I wanted to reach out and ask how the reviews and revisions are going.

I can see that @DamynChipman and @jwallwork23 have opened issues and it looks like @jmark and the team are addressing them. @alecontri have you had a chance to dig into your review yet?

Please let me know if any of you have questions about the process or could use my support in any way.

Thanks all,
Jay
