### 2024-07-20 DLR-AMR/t8code#1161 @jwallwork23
> Please see #1172. I hope i addressed all your concerns. I added further info, linked the Wiki article on Coding guidelines and also updated the Wiki article.

Thanks for addressing this - it's really useful to have this extra detail. Happy for this issue to be closed.
