### 2024-07-23 e5a771c
Switched sc MPI integral types to non-deprecated types
