### 2024-07-24 #6887 @jwallwork23
## @jwallwork23 review summary

I recommend the paper for acceptance, with the condition of the following three remaining checklist items being completed.
> * [ ]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [ ]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

For details on what will be required, see below.

#### Comments on the code

I found t8code to be well-structured and straightforward to get started with.

The installation proceeded as expected on my Windows Subsystem for Linux setup on my laptop running Debian, following the instructions in the [Wiki](https://github.com/DLR-AMR/t8code/wiki/Installation). Although this page is the main reference for installation instructions, it might also be worth mentioning that submodules should be initialised and updated in the `INSTALL` file. That way, users familiar with the standard "configure, make, make install" process can get started.

I ran all the 'general' tutorials, experimenting with different numbers of MPI processes, and viewing and manipulating the outputs in Paraview. I found all the tutorials to work well on my machine.

The automatically generated API documentation is nice. However, there are many data structures whose docstrings are minimal and do not appear in the API docs (`TODO: Document`). I don't think all of these need to be addressed for the paper to be accepted, but it would be worth prioritising doing so in the near future.

I attach a patch file for fixing a few typos in the tutorials: [typos.patch](https://github.com/user-attachments/files/16358935/typos.patch)

#### Comments on the paper

The paper is well written and makes a convincing case in support of t8code. However, there are a few points that should be addressed to improve clarity and present a fair argument by making comparisons with other software, as follows.

> * [ ]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The summary does not adequately introduce mesh adaptation / AMR for a non-specialist audience. (See https://github.com/DLR-AMR/t8code_JOSS/issues/10.)

> * [ ]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

The "Statement of need" section does not adequately state t8code's target audience or how it relates to other work. (See https://github.com/DLR-AMR/t8code_JOSS/issues/11.)

> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

The paper does not compare t8code to other existing packages for mesh adaptation / AMR. (See https://github.com/DLR-AMR/t8code_JOSS/issues/11.)

> * [ ]  **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

I do not have access to a large enough machine to verify reproducibility of the results of the paper (particularly those regarding performance and scaling). As such, I have not been able to mark the above checklist items as completed, nor will I be able to.
