### 2024-07-25 #6887 @elbeejay
Perfect, thanks @jwallwork23 - very much appreciate the specificity of your comments and all of the hyperlinks to connect the comment here to the issues you opened up in the t8code repository.
