### 2024-07-25 DLR-AMR/t8code#1184 @DamynChipman
How to Reproduce Scaling Results
I am working on reproducing some scaling results on `t8code` for a JOSS review: https://github.com/openjournals/joss-reviews/issues/6887.

I noticed the `benchmarks` directory, including the `ExtremeScaling` directory with the bunny example. My first question is this: is that a good "out of the box" problem for me to test the scaling of `t8code` with? If so, I have some questions below on how to run it. If not, what examples would be good candidates to verify the scaling?

For the bunny example, I am running into issues reading in a tetgen file. I have never worked with them, so my issue could simply be lack of experience. I have the `bunny` executable built. In order to run it, I need a tetgen file of the bunny mesh. I got the Stanford bunny mesh data from [here](http://graphics.stanford.edu/data/3Dscanrep/). I copied the file into the `benchmarks/ExtremeScaling` directory and ran the `t8_bunny` example but to no avail.

## Steps to reproduce

I am running this on a MacBook Pro (2021). Once I have it working, I'll repeat the process on the cluster I have access to. Each node of the cluster I will be running on has a 32 core AMD CPU and 4 NVIDIA GPUs. I will only be using the CPUs for this test.

1. Download Stanford bunny zipped file from [here](http://graphics.stanford.edu/data/3Dscanrep/) and unzip into `${downloads}`.

2. Copy the `bun_zipper.ply` into the `ExtremeScaling` directory of `t8code` source (and rename to `bunny`):

```
cd ${t8code-source}/benchmarks/ExtremeScaling
cp ${downloads}/bunny/reconstruction/bun_zipper.ply ./
mv bun_zipper.ply bunny.ply
```

3. Run `bunny` benchmark:

```
mpirun -n 8 ./t8_bunny bunny
```

## Output from above

```
>>> ./t8_bunny bunny
[libsc] This is libsc 2.8.5.406-2b20
[libsc] CPP
[libsc] CPPFLAGS
[libsc] CC                       mpicc
[libsc] CFLAGS                   -g -O2
[libsc] LDFLAGS
[libsc] LIBS                     -lz
[p4est] This is p4est 2.8.6.23-7896
[p4est] CPP
[p4est] CPPFLAGS
[p4est] CC                       mpicc
[p4est] CFLAGS                   -g -O2
[p4est] LDFLAGS
[p4est] LIBS                     -lz
[t8] This is t8 2.0.0.396-758c
[t8] CPP
[t8] CPPFLAGS
[t8] CC                       mpicc
[t8] CFLAGS                   -g -O2
[t8] LDFLAGS
[t8] LIBS                     -lz  -lstdc++
[p4est 0] Failed to open bunny.node
[p4est 0] Failed to read nodes for bunny
[libsc 0] Abort: Failed to read tetgen bunny
[libsc 0] Abort: <unknown>:0
[libsc 0] Abort: Obtained 7 stack frames
[libsc 0] Stack 0: 0   libsc.2.dylib                       0x0000000102766278 sc_abort_handler + 96
[libsc 0] Stack 1: 1   libsc.2.dylib                       0x0000000102766384 sc_abort + 20
[libsc 0] Stack 2: 2   libsc.2.dylib                       0x0000000102765d88 sc_int_compare + 0
[libsc 0] Stack 3: 3   libsc.2.dylib                       0x00000001027663dc sc_abort_collective + 0
[libsc 0] Stack 4: 4   libsc.2.dylib                       0x00000001027671a4 SC_GEN_LOGF + 0
[libsc 0] Stack 5: 5   t8_bunny                            0x00000001022cf81c main + 256
[libsc 0] Stack 6: 6   dyld                                0x0000000189265058 start + 2224
--------------------------------------------------------------------------
MPI_ABORT was invoked on rank 0 in communicator MPI_COMM_WORLD
  Proc: [[46390,0],0]
  Errorcode: 1

NOTE: invoking MPI_ABORT causes Open MPI to kill all MPI processes.
You may or may not see output from other processes, depending on
exactly when Open MPI kills them.
--------------------------------------------------------------------------
```

The issue is the `bunny.ply` file not being the actual file I need. I appear to need a `bunny.node` file. How can I either generate that file from the Stanford bunny data or get that file from elsewhere?
