### 2024-07-26 dc3f475
Update src/t8_cmesh/t8_cmesh_vtk_writer.c
Co-authored-by: Sandro Elsweijer <49643115+sandro-elsweijer@users.noreply.github.com>
