+++ src/t8_cmesh/t8_cmesh_vtk_writer.c
@@ -250,7 +250,8 @@ t8_cmesh_vtk_write_file_ext (t8_cmesh_t cmesh, const char *fileprefix, int write
       /* TODO: We switched to 32 Bit because Paraview could not handle 64 well enough.
        */
       T8_ASSERT (tree->treeid + cmesh->first_tree == (t8_gloidx_t) ((long) tree->treeid + cmesh->first_tree));
-      fprintf (vtufile, " %ld", (long)(tree->treeid + cmesh->first_tree));
+      T8_ASSERT (tree->treeid + cmesh->first_tree ==  tree->treeid + (t8_locidx_t) cmesh->first_tree);
+      fprintf (vtufile, " %ld", tree->treeid + (t8_locidx_t) cmesh->first_tree);
       if (!(sk % 8))
         fprintf (vtufile, "\n         ");
     }
