### 2024-07-29 1bfa267
Add functions in interface
