### 2024-07-29 ae01b68
Update api/t8_fortran_interface/t8_fortran_interface.c
Add header

Co-authored-by: Niklas997 <72195148+Niklas997@users.noreply.github.com>
