### 2024-07-29 2dc44fe
Delete some debugf statements
