### 2024-07-29 f5a7f9d
Update api/t8_fortran_interface/t8_fortran_interface.c
Case if mpi is not used

Co-authored-by: Niklas997 <72195148+Niklas997@users.noreply.github.com>
