### 2024-08-01 62128c7
Update src/t8_geometry/t8_geometry_implementations/t8_geometry_examples.cxx
Co-authored-by: Maël Karembe <82219355+maelk3@users.noreply.github.com>
