### 2024-08-02 3b294c5
Adding triangulated spherical surface cubed version.
