+++ src/t8_cmesh/t8_cmesh_examples.h
@@ -406,6 +406,14 @@ t8_cmesh_new_triangulated_spherical_surface_octahedron (const double radius, sc_
 t8_cmesh_t
 t8_cmesh_new_triangulated_spherical_surface_icosahedron (const double radius, sc_MPI_Comm comm);
 
+/** Construct a triangulated spherical surface of given radius: cube version.
+ * \param [in] radius        Radius of the sphere.
+ * \param [in] comm          The MPI communicator used to commit the cmesh
+ * \return                   A cmesh representing the spherical surface.
+ */
+t8_cmesh_t
+t8_cmesh_new_triangulated_spherical_surface_cube (const double radius, sc_MPI_Comm comm);
+
 /** Construct a quadrangulated spherical surface of given radius.
  * \param [in] radius        Radius of the sphere.
  * \param [in] comm          The MPI communicator used to commit the cmesh
