### 2024-08-05 7a38c35
Updated header file documentation.
