### 2024-08-07 DLR-AMR/t8code#1184 @holke
Hi @DamynChipman
sorry for the late reply, it is holiday season and most of us are currently away.

To be honest, the bunny example is 8+ years old and relies on tetgen which we no longer support. So after seeing your issue we decided to remove it from the code base.

I understand that you want to have some benchmark program that you can run in order to see some scaling results for our core functionality?
If yes, then i suggest the program `./t8_time_forest_partition` from the benchmarks folder.
We use it in several papers, and it is described particularly in https://arxiv.org/pdf/1910.10641 Section 5.2.
In this example we build a mesh geometry, uniformly refine it to a given level l, and then adaptively refine a band of elements, we move this band through the mesh over several time steps.
This example measures `Adapt`  `Partition` `Ghost` and `Balance`, so all of our critical core algorithms.

I recommend using a call like:
`mpirun -np N ./t8_time_forest_partition -g -b -C 0.8 -x -0.4 -X -0.3 -l4 -r3 -O -o -T0.05`

This will build a 1 million element mesh, you can increase the `-l` value if thats to small and runs to quickly (on my machine its ca. 5 seconds on 1 MPI rank).

Here is an overview of the options:
| Option | Description | recommendation |
|---|---|
| -g | Build Ghost layer | don't change|
| -b | 2:1 balance the mesh | don't change |
| -C | CFL number/how fast the mesh moves | don't change |
| -x | Where the band of fine elements starts | decrease for more fine elements |
| -X | Where the band of fine elements stops | increase for more fine elements |
| -l | Uniform refinement level | every steps multiplies # elements by 8 |
| -r | How many refinement levels from the uniform level | change at will |
| -O | Use cylindrical geometry | don't change |
| -o | Do not produce VTK output | Keep it while measuring runtime. For debugging/checking the mesh, leave it out. |
| -T | Simulation end time | Divide by two for each additional uniform level. |

Here is a low dimensional example ("-l 2 -r 3") for the mesh that is created:

![grafik](https://github.com/user-attachments/assets/372fe941-fc4e-40bd-83f6-deb49841a239)

Does this help?

Please let me know if you have further questions.
