### 2024-08-13 #6887 @elbeejay
Fair point and thank you for the reminder @danielskatz.

@jmark can you update us on the status of addressing review comments and the various open issues tied to this JOSS reivew? What is your estimated timeline for making edits and revisions? A number of the issues are relatively small and would only require minor revisions to the manuscript.
