### 2024-08-14 #6887 @jmark
> @jmark can you update us on the status of addressing review comments and the various open issues tied to this JOSS reivew? What is your estimated timeline for making edits and revisions? A number of the issues are relatively small and would only require minor revisions to the manuscript.

Thank you for your reminder! Due to summer vacation the submission got a idle. Most of us are back now and we can look at the issues! Thank you very much for your inputs so far!
