### 2024-08-27 DLR-AMR/t8code#1184 @DamynChipman
Just to confirm, this example requires OpenCascade, correct? After compiling, I ran
```mpirun -np N ./t8_time_forest_partition -g -b -C 0.8 -x -0.4 -X -0.3 -l4 -r3 -O -o -T0.05```
but it stopped and said that example requires OpenCascade. I have been trying to install OpenCascade on my cluster but keep running into system issues not related to `t8code`.
