### 2024-08-29 DLR-AMR/t8code#1184 @jmark
I suggest to use `-L` instead of `-O` even though it is "not recommended" in the table above. This uses a cylinder with linear geometry (linear elements). This does not require t8code to be linked against OpenCascade.
