### 2024-09-02 DLR-AMR/t8code#1184 @DamynChipman
I want to double check to make sure it is running properly locally before moving to a cluster. The benchmark runs to completion and I can mess around with the parameters to increase/decrease the number of elements and view the mesh. However, I don't know if the results are being reported properly. I have used `sc` and `p4est` in the past, including the timing/stats functions. It appears that timing for `adapt`, `ghost`, `partition`, and `balance` are not being accumulated.

I am working on the August 1st commit (https://github.com/DLR-AMR/t8code/commit/62128c743a136a34bd05784597f2065d93a638e4) to avoid a compilation issue reported here (https://github.com/DLR-AMR/t8code/issues/1240). There are no functional differences between `main` and this commit for `benchmarks/time_forest_partition.cxx`.

Here's the output of the following:

```
>>> mpirun -np 8 ./t8_time_forest_partition -g -b -C 0.8 -x -0.4 -X -0.3 -l 4 -r 4 -L -o -T 0.025
```

```
[libsc] This is libsc 2.8.5.999
[p4est] This is p4est 2.8.6.999
[t8] This is t8 2.0.0
[t8] CXX                      /opt/homebrew/bin/mpicxx
[t8] CXXFLAGS                  -O3 -DNDEBUG
[t8] CC                       /opt/homebrew/bin/mpicc
[t8] CFLAGS                    -O3 -DNDEBUG
[t8] LDFLAGS
[t8] LIBS                     P4EST::P4EST SC::SC MPI::MPI_C
[t8] Using delta_t = 0.032000
[t8] Committed cmesh with 4 global trees.
[t8] Start adadpt 0.002101 -0.002101
[t8] Into t8_forest_adapt from 16384 total elements
[t8] Done t8_forest_adapt with 4558320 total elements
[t8] End adadpt 0.199893 0.197792
[t8] Enter  forest partition.
[t8] Start partition 0.199979 0.199979
[t8] End partition 0.208866 0.008887
[t8] Done forest partition.
[t8] Into t8_forest_balance with 4558320 global elements.
[t8] Computed maximum occurring level:	8
[t8] Into t8_forest_ghost with 569790 local elements.
[t8] Start ghost at 0.226103  -0.226103
[t8] End ghost at 0.292601  0.066498
[t8] Done t8_forest_ghost with 569790 local elements and 15485 ghost elements.
[t8] Profiling: 1
[t8] Start adadpt 0.292621 -0.292621
[t8] Into t8_forest_adapt from 4558320 total elements
[t8] Done t8_forest_adapt with 4573216 total elements
[t8] End adadpt 0.354983 0.062362
[t8] Enter  forest partition.
[t8] Start partition 0.355051 0.355051
[t8] End partition 0.361946 0.006895
[t8] Done forest partition.
[t8] Into t8_forest_ghost with 571652 local elements.
[t8] Start ghost at 0.364684  -0.364684
[t8] End ghost at 0.416118  0.051433
[t8] Done t8_forest_ghost with 571652 local elements and 15544 ghost elements.
[t8] Profiling: 1
[t8] Start adadpt 0.416141 -0.416141
[t8] Into t8_forest_adapt from 4573216 total elements
[t8] Done t8_forest_adapt with 4601440 total elements
[t8] End adadpt 0.498088 0.081947
[t8] Enter  forest partition.
[t8] Start partition 0.498677 0.498676
[t8] End partition 0.502745 0.004069
[t8] Done forest partition.
[t8] Into t8_forest_ghost with 575180 local elements.
[t8] Start ghost at 0.505659  -0.505659
[t8] End ghost at 0.557156  0.051497
[t8] Done t8_forest_ghost with 575180 local elements and 15873 ghost elements.
[t8] Profiling: 1
[t8] Start adadpt 0.557175 -0.557175
[t8] Into t8_forest_adapt from 4601440 total elements
[t8] Done t8_forest_adapt with 4648256 total elements
[t8] End adadpt 0.675904 0.118729
[t8] Enter  forest partition.
[t8] Start partition 0.676512 0.676512
[t8] End partition 0.686310 0.009798
[t8] Done forest partition.
[t8] Into t8_forest_ghost with 581032 local elements.
[t8] Start ghost at 0.688951  -0.688951
[t8] End ghost at 0.744801  0.055850
[t8] Done t8_forest_ghost with 581032 local elements and 16422 ghost elements.
[t8] Profiling: 1
[t8] Start adadpt 0.744819 -0.744819
[t8] Into t8_forest_adapt from 4648256 total elements
[t8] Done t8_forest_adapt with 4649376 total elements
[t8] End adadpt 0.863747 0.118927
[t8] Enter  forest partition.
[t8] Start partition 0.864383 0.864383
[t8] End partition 0.869593 0.005210
[t8] Done forest partition.
[t8] Into t8_forest_ghost with 581172 local elements.
[t8] Start ghost at 0.871266  -0.871266
[t8] End ghost at 0.923508  0.052242
[t8] Done t8_forest_ghost with 581172 local elements and 16446 ghost elements.
[t8] Profiling: 1
[t8] Start adadpt 0.923524 -0.923524
[t8] Into t8_forest_adapt from 4649376 total elements
[t8] Done t8_forest_adapt with 4649376 total elements
[t8] End adadpt 1.051267 0.127743
[t8] Done t8_forest_balance with 4649376 global elements.
[t8] Statistics for   forest balance: Adapt time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0646732 (0.003 = 4.64%)
[t8]    Minimum attained at rank       3: 0.062135
[t8]    Maximum attained at rank       1: 0.068688
[t8] Statistics for   forest balance: Adapt time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0821344 (0.000428 = 0.521%)
[t8]    Minimum attained at rank       3: 0.081806
[t8]    Maximum attained at rank       6: 0.083232
[t8] Statistics for   forest balance: Adapt time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.118677 (0.00042 = 0.354%)
[t8]    Minimum attained at rank       2: 0.118279
[t8]    Maximum attained at rank       6: 0.119743
[t8] Statistics for   forest balance: Adapt time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.119126 (0.00111 = 0.931%)
[t8]    Minimum attained at rank       2: 0.118523
[t8]    Maximum attained at rank       6: 0.122047
[t8] Statistics for   forest balance: Adapt time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.127708 (0.000561 = 0.439%)
[t8]    Minimum attained at rank       2: 0.127265
[t8]    Maximum attained at rank       6: 0.129152
[t8] Statistics for   forest balance: Total adapt time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.512319 (0.00478 = 0.933%)
[t8]    Minimum attained at rank       2: 0.508419
[t8]    Maximum attained at rank       6: 0.522614
[t8] Summary = [ 0.0646732 0.0821344 0.118677 0.119126 0.127708 0.512319 ];
[t8] Maximum = [ 0.068688 0.083232 0.119743 0.122047 0.129152 0.522614 ];
[t8] Statistics for   forest balance: Ghost time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0512449 (0.000435 = 0.848%)
[t8]    Minimum attained at rank       6: 0.050128
[t8]    Maximum attained at rank       3: 0.051583
[t8] Statistics for   forest balance: Ghost time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0515504 (0.000421 = 0.816%)
[t8]    Minimum attained at rank       6: 0.050481
[t8]    Maximum attained at rank       2: 0.051946
[t8] Statistics for   forest balance: Ghost time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0556621 (0.00111 = 2%)
[t8]    Minimum attained at rank       6: 0.052731
[t8]    Maximum attained at rank       2: 0.056257
[t8] Statistics for   forest balance: Ghost time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0522777 (0.000561 = 1.07%)
[t8]    Minimum attained at rank       6: 0.050832
[t8]    Maximum attained at rank       2: 0.05272
[t8] Statistics for   forest balance: Total ghost time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.210735 (0.00251 = 1.19%)
[t8]    Minimum attained at rank       6: 0.204172
[t8]    Maximum attained at rank       2: 0.212393
[t8] Summary = [ 0.0512449 0.0515504 0.0556621 0.0522777 0.210735 ];
[t8] Maximum = [ 0.051583 0.051946 0.056257 0.05272 0.212393 ];
[t8] Statistics for   forest balance: Partition time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.00686963 (0.00138 = 20%)
[t8]    Minimum attained at rank       3: 0.005146
[t8]    Maximum attained at rank       1: 0.009283
[t8] Statistics for   forest balance: Partition time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.00521925 (0.000772 = 14.8%)
[t8]    Minimum attained at rank       0: 0.004069
[t8]    Maximum attained at rank       2: 0.00628
[t8] Statistics for   forest balance: Partition time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.00729638 (0.00281 = 38.5%)
[t8]    Minimum attained at rank       7: 0.004249
[t8]    Maximum attained at rank       2: 0.011448
[t8] Statistics for   forest balance: Partition time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.00519713 (0.000547 = 10.5%)
[t8]    Minimum attained at rank       5: 0.004715
[t8]    Maximum attained at rank       2: 0.006559
[t8] Statistics for   forest balance: Total partition time
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0245824 (0.00407 = 16.6%)
[t8]    Minimum attained at rank       5: 0.0211
[t8]    Maximum attained at rank       1: 0.031333
[t8] Summary = [ 0.00686963 0.00521925 0.00729638 0.00519713 0.0245824 ];
[t8] Maximum = [ 0.009283 0.00628 0.011448 0.006559 0.031333 ];
[t8] Into t8_forest_ghost with 581172 local elements.
[t8] Start ghost at 1.055855  -1.055855
[t8] End ghost at 1.108176  0.052320
[t8] Done t8_forest_ghost with 581172 local elements and 16446 ghost elements.
[t8] Printing stats for cmesh.
[t8] Statistics for   cmesh: Number of trees sent.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: Number of ghosts sent.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: Number of trees received.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: Number of ghosts received.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: Number of bytes sent.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: Number of processes sent to.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: First tree is shared.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           -8 (0 = 0%)
[t8]    Minimum attained at rank       0: -8
[t8]    Maximum attained at rank       0: -8
[t8] Statistics for   cmesh: Partition runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   cmesh: Commit runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           3.1e-05 (2.47e-05 = 79.7%)
[t8]    Minimum attained at rank       6: 4e-06
[t8]    Maximum attained at rank       0: 8e-05
[t8] Statistics for   cmesh: Number of geometry evaluations.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           651125 (7.44e+05 = 114%)
[t8]    Minimum attained at rank       0: 2304
[t8]    Maximum attained at rank       4: 1.81606e+06
[t8] Statistics for   cmesh: Accumulated geometry evaluation runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0189933 (0.0216 = 114%)
[t8]    Minimum attained at rank       6: 8e-05
[t8]    Maximum attained at rank       5: 0.052864
[t8] Summary = [ 0 0 0 0 0 0 -8 0 3.1e-05 651125 0.0189933 ];
[t8] Maximum = [ 0 0 0 0 0 0 -8 0 8e-05 1.81606e+06 0.052864 ];
[t8] Printing stats for forest.
[t8] Statistics for   forest: Number of elements sent.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           441347 (4.62e+05 = 105%)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       5: 1.13548e+06
[t8] Statistics for   forest: Number of elements received.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           441347 (2.23e+05 = 50.6%)
[t8]    Minimum attained at rank       4: 0
[t8]    Maximum attained at rank       1: 569790
[t8] Statistics for   forest: Number of bytes sent.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           1.05924e+07 (1.11e+07 = 105%)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       5: 2.72517e+07
[t8] Statistics for   forest: Number of processes sent to.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           1.375 (0.992 = 72.2%)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       4: 3
[t8] Statistics for   forest: Number of ghost elements sent.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           33532.8 (6.59e+03 = 19.6%)
[t8]    Minimum attained at rank       0: 18758
[t8]    Maximum attained at rank       6: 39696
[t8] Statistics for   forest: Number of ghost elements received.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           33532.8 (7.09e+03 = 21.1%)
[t8]    Minimum attained at rank       0: 16446
[t8]    Maximum attained at rank       4: 39682
[t8] Statistics for   forest: Number of processes we sent ghosts to/received from.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   forest: Adapt runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   forest: Partition runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0146836 (0.00525 = 35.8%)
[t8]    Minimum attained at rank       1: 0.008687
[t8]    Maximum attained at rank       3: 0.023082
[t8] Statistics for   forest: Commit runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.855875 (1.79e-06 = 0.000209%)
[t8]    Minimum attained at rank       3: 0.855872
[t8]    Maximum attained at rank       4: 0.855877
[t8] Statistics for   forest: Ghost runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.0523034 (0.00076 = 1.45%)
[t8]    Minimum attained at rank       6: 0.050317
[t8]    Maximum attained at rank       2: 0.052796
[t8] Statistics for   forest: Ghost waittime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   forest: Balance runtime.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0.831332 (0.000132 = 0.0159%)
[t8]    Minimum attained at rank       1: 0.831124
[t8]    Maximum attained at rank       2: 0.831623
[t8] Statistics for   forest: Balance rounds.
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           5 (0 = 0%)
[t8]    Minimum attained at rank       0: 5
[t8]    Maximum attained at rank       0: 5
[t8] Summary = [ 441347 441347 1.05924e+07 1.375 33532.8 33532.8 0 0 0.0146836 0.855875 0.0523034 0 0.831332 5 ];
[t8] Maximum = [ 1.13548e+06 569790 2.72517e+07 3 39696 39682 0 0 0.023082 0.855877 0.052796 0 0.831623 5 ];
[t8] Statistics for   new
[t8]    Global number of values:      16
[t8]    Mean value (std. dev.):           0.00203125 (3.23e-05 = 1.59%)
[t8]    Minimum attained at rank       0: 0.002
[t8]    Maximum attained at rank       5: 0.002106
[t8] Statistics for   adapt
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   ghost
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   partition
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   balance
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           0 (0)
[t8]    Minimum attained at rank       0: 0
[t8]    Maximum attained at rank       0: 0
[t8] Statistics for   total
[t8]    Global number of values:       8
[t8]    Mean value (std. dev.):           1.10902 (0.000273 = 0.0246%)
[t8]    Minimum attained at rank       1: 1.10871
[t8]    Maximum attained at rank       5: 1.10947
[t8] Summary = [ 0.00203125 0 0 0 0 1.10902 ];
[t8] Maximum = [ 0.002106 0 0 0 0 1.10947 ];
```

Am I right in seeing that the stats are not being accumulated and reported? Or are they being reported in the `forest balance: <Adapt, Ghost, Partition> time` sections? What needs to change in order to accumulate/report the timing for these stages?
