### 2024-09-04 DLR-AMR/t8code#1184 @DamynChipman
Just confirming that PR https://github.com/DLR-AMR/t8code/pull/1242 fixed the compilation issue and I have reran the same commands as above with `main` with the same results. I want to make sure I get the right timing results for this benchmark to correctly represent and reproduce the scalability of `t8code`.
