### 2024-09-06 #6887 @jmark
> @jmark looks like you are chipping away at the issues opened up in the review process - please continue to do so and reach out to the reviewers to close issues as needed as you resolve them. Let me know if there's anything you need from me at this time.

I've been occupied by other tasks on short notice. I'm sorry to prolong the refactoring process longer as necessary. I allocated today to process the issues. I thank you very much for you patience.
