### 2024-09-09 #6887 @DamynChipman
The `t8code` team fixed a couple issues I raised that popped up. My last open issue is the one regarding the reproducibility of the scaling results on a cluster. I have ran small scale timing on a provided benchmark on my laptop. The software and benchmark are installed on the cluster; I will submit that job tomorrow and expect to have it run by the end of the week.
