### 2024-09-13 #6887 @jwallwork23
Thank you to the authors for responding to the issues raised as part of my review. I am happy to approve from my side.

I did notice a few typos in the current version of the manuscript that it would be good to fix, though:
* L56: "then" should be "than"
* L58: "till" should be "until"
* L63-L64: Suggest rewriting sentence as "The starting point for usage of t8code is a coarse, unstructured, conformal input mesh."
* L96: "mesh operation" should be "mesh operations".

Also, would the authors be willing to use a logarithmic scale on the vertical axis in Figure 3? This would seem more suitable.
