### 2024-09-17 #6887 @elbeejay
@alecontri - do you mind taking a second pass at this review when you get a chance? It seems like @jmark and colleagues have made a number of revisions since you tested the software for JOSS the first time.
