### 2024-09-18 #6887 @alecontri
> @alecontri - do you mind taking a second pass at this review when you get a chance? It seems like @jmark and colleagues have made a number of revisions since you tested the software for JOSS the first time.

Yes, absolutely, I will go through it again in the weekend. Thanks for the reminder
