### 2024-09-22 #6887 @alecontri
## Review summary @alecontri

I thank the authors for having assessed the majority of the issues I raised.

I find the article well written. It gives a balanced presentation of t8code and the working landscape in which its impact is of relevance.

I was able, following the provided installation guide on the [Wiki](https://github.com/DLR-AMR/t8code/wiki/Installation), to install t8code on my Apple machine running Sonoma on a M1 processor. The automated tests compiled correctly and the Tutorials provided enough information to begin to interface with the software and to understand the building blocks of it.

I agree with previous comments on how there are some uncompleted docstrings that should be completed in the near future but do not prevent, from my side, the article from being accepted.

I thus recommend the paper for acceptance, with the condition that the following issue is handled:
https://github.com/DLR-AMR/t8code_JOSS/issues/14#issue-2423104254

Regarding the checklist:

### Functionality

I do not have access to an infrastructure so to reproduce the statements done on the reproducibility, performance and scalability of the code. For this only reason, I won't be able to mark the above checklist items as completed, either now or in the future.

Finally: in line 64 maybe you meant "which we denote as coarse mesh" instead of "which we denote a coarse mesh"?
