### 2024-09-24 #6887 @elbeejay
@DamynChipman can you provide us with an update on your last outstanding issue?

> My last open issue is the one regarding the reproducibility of the scaling results on a cluster. I have ran small scale timing on a provided benchmark on my laptop. The software and benchmark are installed on the cluster; I will submit that job tomorrow and expect to have it run by the end of the week.

---

@alecontri it sounds like your issue https://github.com/DLR-AMR/t8code_JOSS/issues/14#issue-2423104254 as well as these last two minor reviewer checklist items still need to be resolved. Is that correct? @jmark you'd mentioned in the issue that you'd check if you had relevant data (comparisons/charts/graphs to other similar software) - have you made any progress there?
