### 2024-09-24 #6887 @DamynChipman
I have ran scaling results up to 32 MPI ranks with good strong scaling. Beyond 32 ranks the runs quit on running out of memory. I have raised this concern on my open issue (debugging if it is the benchmark or configuration of my system/job).

The remaining items to check off of my review relate to the reproducibility of the scalability. If strong scaling up to 32 cores is satisfactory for this review, I am good to check my remaining items off.
