### 2024-09-27 DLR-AMR/t8code#1184 @Davknapp
Hey @DamynChipman ,
can you provide us the input parameters for this run, to reproduce it?
In general you shouldn't have any problems running this example on a cluster, we have tested it using much more than 32 procs. But this has been a couple of years ago, so we will double check again, especially with your input parameters.

One part of your output might indicate that you just ran out of memory:
```
[libsc 31] Abort: Returned NULL from malloc
[libsc 31] Abort: /home/dchipman1/packages/t8code/sc/src/sc.c:398
```
Maybe your parameter-combination produced a mesh that was to large to handle for the machine?
