### 2024-10-08 #6887 @elbeejay
@jmark where are we with the comparisons to other algorithsm?

@DamynChipman were you able to finish your scaling/performance verification?

Thanks all
