### 2024-10-10 DLR-AMR/t8code#1184 @jmark
@DamynChipman I tried to run this benchmark example on my machine with 64 ranks. It runs just fine.
```bash
mpirun -n 64 ~/install/t8code/main/bin/t8_time_forest_partition -g -b -C 0.8 -x -0.4 -X -0.3 -l 4 -r 5 -L -o -T 0.025
```

I use `MPICH version 4.0.2` and `gcc version 12.1.0` on a `Linux Kernel 6.8.0-45-generic` and `22.04.1-Ubuntu`.
