### 2024-10-24 DLR-AMR/t8code#1184 @DamynChipman
# Large (ish) Scale Results

Machine: Falcon (Dual Intel Xeon 18 core nodes)
Command:
```
mpirun -n $N ./t8_time_forest_partition -g -b -C 0.8 -x -0.2 -X 0.2 -l 4 -r 4 -L -o -T 0.025 >> output_n$N.txt
```

Results:

|        |1       |2       |4       |8       |16      |32      |64       |128      |256      |512       |
|--------------|--------|--------|--------|--------|--------|--------|---------|---------|---------|----------|
|Adapt [sec]         |22.198  |15.3656 |4.88433 |3.51135 |2.74351 |1.75711 |1.25502  |0.832045 |0.963431 |2.83504   |
|Ghost [sec]         |0       |4.92692 |2.57878 |2.4864  |1.56907 |0.939954|0.627435 |0.385915 |0.289718 |0.207101  |
|Partition [sec]     |3.50545 |1.78216 |0.728757|0.433819|0.198721|0.152427|0.0674145|0.0380946|0.0197075|0.00966559|
|Total [sec]         |45.7854 |35.9366 |12.8784 |10.5972 |6.90435 |4.21385 |3.01467  |2.16236  |3.01856  |4.97634   |
|Total Elements|34723392|34723392|34723392|34723392|34723392|34723392|34723392 |34723392 |34723392 |34723392  |
---

# Comments

Turns out the issue was just the machine I was running on (it's a rather picky cluster that is currently being worked on as a decommissioned national lab computer). I was able to run up to 512 MPI ranks with runs beyond that failing due to the machine configuration. For the purposes of reproducing this benchmark, I am satisfied and impressed with the speed and memory footprint for over 30M elements!

Thank you for your guidance on reproducing this result!
