### 2024-10-24 #6887 @DamynChipman
My reproduced scaling results can be found here: https://github.com/DLR-AMR/t8code/issues/1184. Due to machine limitations that are beyond my control, I am unable to run to higher core count. However, the `t8code` benchmark I ran shows good strong scaling. I am satisfied and impressed with the size of the problem and memory footprint for a benchmark that included over 30M elements at the finest AMR levels.

I have completed all items in my review checklist. Working with the `t8code` team over GitHub issues was efficient and productive. The paper is well-written and the caliber of the software is exactly what open source software should be aiming for. Installing `t8code` on my laptop and on clusters was straightforward. The test cases provide a potential user confidence that the code is installed correctly and the examples are great starting points for expansion and applications. I recommend `t8code` be accepted.

I appreciate the chance to review and your patience working through the scaling results.
