### 2024-10-24 #6887 @jmark
> @jmark - I'd like us to move towards some sort of resolution of the reviewers' valid concerns about scalability and comparisons of `t8code` to other algorithms. Are you able to run any of the tests needed to assuage these concerns at your institution?

I'm working on it.
