### 2024-10-30 #6887 @jmark
> > @jmark - I'd like us to move towards some sort of resolution of the reviewers' valid concerns about scalability and comparisons of `t8code` to other algorithms. Are you able to run any of the tests needed to assuage these concerns at your institution?
>
> I'm working on it.

Short status update.

Intermediate results are available. I created a PR for this on t8code side: https://github.com/DLR-AMR/t8code/pull/1287

The plan was to compare ghost layer creation time between t8code and p4est. Of course, only quad/hex meshes can be compared. So far, t8codes ghost layer creation is consistently around 1,5 times slower. This is expected. The algorithm is more flexible and more powerful since it supports multiple element types. p4est's algorithms on the other side, are optimized for quad/hex elements. Nevertheless, a pretty decent result. In the next days, I will add more detailed results into the JOSS paper.
