### 2024-11-11 #6887 @jmark
I did some scaling tests wit t8code and compared them with p4est. See https://github.com/DLR-AMR/t8code_JOSS/issues/14.

I added the results to the JOSS paper: https://github.com/DLR-AMR/t8code_JOSS/blob/main/paper.pdf. Specifically, in the **Performance** section.

Let me know if you agree and if there is anything else I can do.
