### 2024-11-14 #6887 @elbeejay
@jmark we will need the updated paper markdown file to exist somewhere in the primary `t8code` repository, in a branch is fine, just let me know which one to point the editorial bot at, thanks!
