### 2024-11-14 #6887 @jmark
> @jmark we will need the updated paper markdown file to exist somewhere in the primary `t8code` repository, in a branch is fine, just let me know which one to point the editorial bot at, thanks!

Ups, sorry. I converted the `JOSS` directory in the `JOSS-paper` branch to a git submodule linking to the `t8code_JOSS` repository.
I reverted back to the old version with the files living directly in a t8code branch.

The bot should now be able to find the files with its existing configuration. Here for reference:

https://github.com/DLR-AMR/t8code/tree/JOSS-paper/JOSS

Linking the paper sources as `git submodule` would be a nice option in my opinion.
