### 2024-11-15 #6887 @elbeejay
> Linking the paper sources as git submodule would be a nice option in my opinion.

I'll share this feedback with the JOSS team!
