### 2024-11-18 #6887 @jmark
> > @jwallwork23 - I think you finished your review (with caveats around what you were not able to test/verify); do you mind double-checking your reviewer checklist above to make sure it is as complete as you are comfortable making it?
>
> Yes, I can confirm that it's complete. I've read through the latest version of the manuscript and it looks great, except a typo "terraybyte" in the caption for Figure 3 (compared with "terrabyte" in the Performance section).

Good catch! Fixed.
