### 2024-11-20 #6887 @elbeejay
A few notes besides the suggestion to double-check for DOIs for the references flagged above:

- Please remove the references the paper makes to itself (e.g., line 43); in line 49 please feel free to refer the reader to the documentation/main web page like you do in the Summary section
- Line 77: Consider putting the reference to Figure 2 in parentheses "(see Figure 2)"
- Same suggestion as above for lines 99-100 which reference Figure 3

In addition to these notes I'd ask that you please provide me with the DOI for the archived version of the software to accompany this publication, as well as an updated software version release number (v3.0.0 now?), in addition to confirming that you've checked authors/affiliations and licenses per the post-review checklist above. Thanks!
