### 2024-11-20 #6887 @jmark
@elbeejay Thank you very much for your proof-reading efforts and your suggestions.

---

>     * Please remove the references the paper makes to itself (e.g., line 43); in line 49 please feel free to refer the reader to the documentation/main web page like you do in the Summary section

This is actually a reference to a short note from around 1.5 years ago published at the International Meshing Round Table 2023 symposium affiliated to the SIAM CSE Amsterdam. It was about t8code v1.0.0. But I get your point and removed the reference from the JOSS paper.

---

>     * Line 77: Consider putting the reference to Figure 2 in parentheses "(see Figure 2)"

For the sake of good writing style I reformulated this into a proper sentence.

---

>     * Same suggestion as above for lines 99-100 which reference Figure 3

Same as above.

---

> In addition to these notes I'd ask that you please provide me with the DOI for the archived version of the software to accompany this publication, as well as an updated software version release number (v3.0.0 now?),

The DOI of t8code is https://doi.org/10.5281/zenodo.7034838 .

The latest version of t8code at the time of writing is `v3.0.0`.

---

> in addition to confirming that you've checked authors/affiliations and licenses per the post-review checklist above. Thanks!

I have question about this which I ask in a separate comment.
