### 2024-11-20 #6887 @jmark
The post-review checklist says:
> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

The Zenodo DOI lists all authors who have in any way contributed to t8code. And be it just one line of code. Is it really necessary to list the complete list in the DOI in the JOSS paper. Can it be limited to the main developers with significant contributions to the code?
