### 2024-11-21 #6887 @elbeejay
> The post-review checklist says:
>
> > Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>
> The Zenodo DOI lists all authors who have in any way contributed to t8code. And be it just one line of code. Is it really necessary to list the complete list in the DOI in the JOSS paper? Can it be limited to the main developers with significant contributions to the code?

Hi @jmark - you should be able to edit the Zenodo page to match the author list tied to the JOSS publication. The JOSS publication can be limited to the primary developers with significant contributions to the project.
