### 2024-11-22 #6887 @jmark
@elbeejay

Following changes to the paper were made:
* Updated the authors list to match it with the Zenodo DOI link.
* Replaced `LICENSE` file with a copy from the t8code repository so they match exactly. License is GPLv2 now.

This should close the last open checkbox. Let me know if there is anything else to do.
