### 2024-11-25 #6887 @elbeejay
Great, thanks @jmark.

@danielskatz - do we require the order of the authors in the zenodo archive to match the order in the JOSS paper?
