### 2024-11-26 #6887 @jmark
> > It probably should, unless there's a good reason for it not to.
>
> Ok, I see. I report back to my co-authors and discuss how we want to properly match the authors list at Zenodo with the JOSS paper.

The latest Zenodo authors list was compiled solely based on Github metrics (#commits, #changed lines, etc.) specifically contributing to t8code v3.0.0 and is not supposed to be changed afterwards.

I kindly ask you to accept the author list of the current JOSS paper draft as it is which better reflects the contribution to the paper and to the code.

Alternatively, I can offer you to change the Zenodo DOI to t8code v1.0.0 in which the given author list reflects the main contributors to t8code back in 2022 (and also today). If you consent I will match the paper draft with this.

t8code v1.0.0: https://zenodo.org/records/7104580
