### 2024-11-26 #6887 @danielskatz
> I kindly ask you to accept the author list of the current JOSS paper draft as it is which better reflects the contribution to the paper and to the code.

yes, this is what we would prefer the zenodo author list be changed to.

You can change the zenodo metadata manually. It defaults to something based on the repo, but this is just the default, not what you need to accept. You can edit the metadata and republish the entry without a new DOI being generated.
