### 2024-12-12 #6887 @jmark
> @jmark - were you able to resolve the zenodo/paper author list conflict? I think I checked last week and it looked like things were still mis-aligned. Thanks!

I'm working towards a resolution.
