### 2024-12-13 #6887 @jmark
@elbeejay We came to a resolution which hopefully pleases all sides!

Please update to the following DOI: https://doi.org/10.5281/zenodo.14418226

This matches with the author's list in the JOSS paper.
