### 2024-12-14 #6887 @elbeejay
@jmark great, thanks. This generally works for me, I will note that I think you've missed including Ole Albers' ORCID ID in the JOSS paper itself.
