### 2024-12-14 #6887 @elbeejay
@jmark is it possible to revise the title of the Zenodo entry to match the JOSS paper title exactly? "t8code - modular adaptive mesh refinement in the exascale era"

No problem with the caveats and other citation guidance you provide people regarding the distinction between this archive for the JOSS paper and the more general one.
