### 2024-12-16 #6887 @jmark
> @jmark is it possible to revise the title of the Zenodo entry to match the JOSS paper title exactly? "t8code - modular adaptive mesh refinement in the exascale era"
>
> No problem with the caveats and other citation guidance you provide people regarding the distinction between this archive for the JOSS paper and the more general one.

This should be doable. And thank you for the hint of Ole Albers missing ORCID id. I do not have write access to the Zenodo link. My colleague, who has write access, is on Christmas vacation. So this probably will have to wait till beginning next year. I'll write him a note, though.

I again thank you and your team very much for you input and your effort in reviewing and proof reading.

I wish you all some peaceful and nice holidays and a good start into the new year!
