### 2025-01-06 #6887 @jmark
> @jmark is it possible to revise the title of the Zenodo entry to match the JOSS paper title exactly? "t8code - modular adaptive mesh refinement in the exascale era"

The revision is done.

Let me know if there is anything else left to do.
