### 2025-01-06 #6887 @elbeejay
I think we've managed to resolve the last few outstanding issues surrounding the titles and author lists.
