### 2025-01-08 #6887 @jmark
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/12641921185?check_suite_focus=true). The generated XML metadata file is invalid.

Not sure what to make of this. Does this has anything to do with non-conforming bib entries?
