### 2025-01-08 #6887 @danielskatz
It looks like there is a DOI field that doesn't contain a valid DOI.  Where is the paper source?
