### 2025-01-09 #6887 @jmark
> Ah, I found it. The DOI in
>
> > @PhDThesis{holke_scalable_2018,
> > type = {PhD thesis},
> > title = {Scalable algorithms for parallel tree-based adaptive mesh refinement with general element types},
> > school = {Rheinische Friedrich-Wilhelms-Universität Bonn},
> > author = {Holke, Johannes},
> > year = {2018},
> > doi = {20.500.11811/7661}
> > }
>
> Is not a DOI but a handle, which the DOI resolver can deal with, but isn't a DOI. Please change this is a URL field with the full URL.

I hope this commit does the trick: https://github.com/DLR-AMR/t8code/commit/9884ce8a8417b85b65201f57b86c3a2ac7a38020
