### 2025-01-09 #6887 @elbeejay
Hm, I don't know exactly which entry is causing this problem, but some atypical patterns in your bib entries include:

- [`note = {\url{https://doi.org/10.1007/s00366-006-0049-3}}`](https://github.com/DLR-AMR/t8code/blob/JOSS-paper/JOSS/paper.bib#L574) - perhaps this can be changed to just a `url` field rather than a `note`
- [`howpublished={\url{https://github.com/trixi-framework/Trixi.jl}},`](https://github.com/DLR-AMR/t8code/blob/JOSS-paper/JOSS/paper.bib#L550) - this too could be changed to a `url` field

Do you have any other suggestions @danielskatz? The entry you'd previous flagged has been revised to:

```
@phdthesis{holke_scalable_2018,
  type = {PhD thesis},
  title = {Scalable algorithms for parallel tree-based adaptive mesh refinement with general element types},
  school = {Rheinische Friedrich-Wilhelms-Universität Bonn},
  author = {Holke, Johannes},
  year = {2018},
  url = {https://hdl.handle.net/20.500.11811/7661}
}
```
