### 2025-01-13 #6887 @jmark
I cleaned up the bibtex file with a linter. Based on that I did some changes here and there. Would you try again and see if this fixes the errors?
