### 2025-01-29 #6887 @xuanxu
In the metadata of the paper.md file, there are cases of author information added using invalid syntax:
[Some of the authors names](https://github.com/DLR-AMR/t8code/blob/JOSS-paper/JOSS/paper.md?plain=1#L25-L66) are using only the `given-name` property for the complete name without providing a surname or family name. They should be changed to use the `name` property or a complete set of components.
[Here are the relevant docs](https://joss.readthedocs.io/en/latest/paper.html#names)
