### 2025-01-31 #6887 @jmark
@xuanxu Thank you for pointing that out! It is fixed now: https://github.com/DLR-AMR/t8code/commit/059a017241aa34ebc396939ed53c14e241ffa9cb

@elbeejay I also added the missing DOI: https://github.com/DLR-AMR/t8code/commit/2c12db1a6388cdc6025476c8becd44d5eb515e18
