### 2025-02-01 #6887 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
