### 2025-02-02 #6887 @danielskatz
👋 @jmark - some of the affiliations in your paper don't have countries. Please add them.

Additionally, I've suggested a number of small changes in https://github.com/DLR-AMR/t8code/pull/1369. Please merge this or let me know what you disagree with.
