### 2025-02-03 #6887 @jmark
> 👋 [@jmark](https://github.com/jmark) - some of the affiliations in your paper don't have countries. Please add them.
>
> Additionally, I've suggested a number of small changes in [DLR-AMR/t8code#1369](https://github.com/DLR-AMR/t8code/pull/1369). Please merge this or let me know what you disagree with.

@danielskatz Thank you very much for taking your time proof-reading the submission! Your proposed corrections have been merged.
