### 2025-02-03 #6887 @elbeejay
@danielskatz, due to HPC access limitations, our reviewers were not all able to sign off on every element of their checklists. Caveats were made, I'll link and quote the appropriate comments below:

https://github.com/openjournals/joss-reviews/issues/6887#issuecomment-2247225640

> I do not have access to an infrastructure so to reproduce the statements done on the reproducibility, performance and scalability of the code. For this only reason, I won't be able to mark the above checklist items as completed, either now or in the future.

and

https://github.com/openjournals/joss-reviews/issues/6887#issuecomment-2483702086

> The only item not checked in my list is thus the scaling for lager number of processors (Performance), that has already been taken care of and that I unfortunately cannot confirm with my available means.

Due to those caveats there are a few (intentionally) unchecked boxes in reviewer checklists.
