### 2025-02-04 #6887 @danielskatz
@jmark - sorry, I missed one bib entry the first time. Can you also merge https://github.com/DLR-AMR/t8code/pull/1377 ?

And please check the proof above to make sure everything else looks ok...
