### 2025-02-06 #6887 @jmark
> [@jmark](https://github.com/jmark) - sorry, I missed one bib entry the first time. Can you also merge [DLR-AMR/t8code#1377](https://github.com/DLR-AMR/t8code/pull/1377) ?
>
> And please check the proof above to make sure everything else looks ok...

@danielskatz Done!
